[Chore] Split asset_execution.rs into nine submodules
One 4 514-line file held a 2 376-line `impl AssetExecutionContext` with
eighty methods, spanning provider calls, the bundle cache, downloading,
task planning, the 3D pipeline and the run loop, plus twenty-five free
functions and 1 500 lines of tests. Rust allows an inherent impl to be
split across modules of the same crate, so each group moves to its own
file and the type keeps every method.

    model     the types the path moves around      (leaf)
    crypto    deobfuscation                        -> model
    planning  filters and asset info -> tasks      -> model
    progress  the update enum and its sender       -> model
    provider  cookies, asset info, bundle URLs     -> crypto model
    cache     the persistent bundle cache          -> crypto model planning
    runner    scheduling, recording, finishing     -> model progress
    haruki_3d the 3D export pipeline               -> model planning progress
    download  fetching and writing payloads        -> the rest

A cycle check over the `use super::` edges reports none.

Pure move: bodies are byte-for-byte, and an item-by-item comparison
against the original counts 114 in and 114 out. The only edits are
visibility -- methods and struct fields that were private inside one file
are `pub(super)` so siblings can still reach them -- and imports.

`fetch_live_asset_bundle_info`, `should_download_bundle` and
`deobfuscate` were `pub` in the flat module and so reachable from outside
it; the facade re-exports them to keep that unchanged. The first has no
caller anywhere, which the flat file could not show. Removing it is a
change in behaviour surface, not a move, so it is left for later.
