+++ src/core/asset_execution.rs
@@ -1,2978 +1,26 @@
-use std::collections::{HashMap, HashSet};
-use std::net::{IpAddr, Ipv4Addr};
-use std::path::{Component, Path, PathBuf};
-use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-use std::sync::Arc;
-use std::time::{Duration, Instant};
+//! Running an asset update: planning, downloading, exporting, recording.
+//!
+//! This file is the facade; the submodules below hold the implementation.
+//! Every external path -- `crate::core::asset_execution::Thing` -- is
+//! unchanged.
 
-use cbc::cipher::{block_padding::Pkcs7, BlockModeDecrypt, KeyIvInit};
-use chrono::FixedOffset;
-use reqwest::header::{
-    HeaderMap, HeaderValue, ACCEPT, ACCEPT_ENCODING, ACCEPT_LANGUAGE, COOKIE, SET_COOKIE,
-    USER_AGENT,
+mod cache;
+mod crypto;
+mod download;
+mod haruki_3d;
+mod model;
+mod planning;
+mod progress;
+mod provider;
+mod runner;
+
+pub use crypto::{decrypt_asset_bundle_info, deobfuscate};
+pub use model::{
+    AssetBundleDetail, AssetBundleInfo, AssetCategory, AssetExecutionContext, Haruki3dExportSummary,
 };
-use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
-use tokio::sync::mpsc::UnboundedSender;
-use tokio::sync::{OwnedSemaphorePermit, Semaphore};
-use tokio::task::JoinSet;
-
-use crate::core::cleanup::remove_file_if_exists;
-use crate::core::config::{AppConfig, AssetHttpVersion, RegionConfig, RegionProviderConfig};
-use crate::core::download_records::{load_download_record, save_download_record, DownloadRecord};
-use crate::core::errors::{format_reqwest_error_chain, AssetExecutionError};
-use crate::core::export_pipeline::{
-    export_unity_asset_bundle_payloads_with_registry, post_process_exported_files,
-    NativeObjectReadPlanStats, NativeSemanticExportPathRegistry, UnityAssetBundlePayloadExport,
-};
-use crate::core::git_sync::sync_chart_hashes;
-use crate::core::models::{AssetUpdateRequest, ExecutionSummary, JobPhase};
-use crate::core::regions::{compile_patterns, first_match_index, matches_any};
-use crate::core::retry::retry_async;
-
-type Aes128CbcDec = cbc::Decryptor<aes::Aes128>;
-type Aes192CbcDec = cbc::Decryptor<aes::Aes192>;
-type Aes256CbcDec = cbc::Decryptor<aes::Aes256>;
-static BUNDLE_CACHE_WRITE_SEQUENCE: AtomicUsize = AtomicUsize::new(0);
-
-fn asset_category_name(category: &AssetCategory) -> &'static str {
-    match category {
-        AssetCategory::StartApp => "StartApp",
-        AssetCategory::OnDemand | AssetCategory::LivePv | AssetCategory::Other(_) => "OnDemand",
-    }
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-pub enum AssetCategory {
-    StartApp,
-    OnDemand,
-    LivePv,
-    Other(String),
-}
-
-impl<'de> Deserialize<'de> for AssetCategory {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        // Treat nil/null as Other("") — matches Go's zero-value coercion.
-        let raw = Option::<String>::deserialize(deserializer)?.unwrap_or_default();
-        Ok(match raw.as_str() {
-            "StartApp" | "startApp" => Self::StartApp,
-            "OnDemand" | "onDemand" => Self::OnDemand,
-            "Live_pv" | "live_pv" | "LivePv" | "livePv" => Self::LivePv,
-            other => Self::Other(other.to_string()),
-        })
-    }
-}
-
-/// Deserializes a msgpack/JSON null or missing value as an empty String.
-/// Go silently coerces nil → zero value for non-pointer types; this helper
-/// mirrors that behavior for String fields.
-fn de_null_as_empty_string<'de, D>(deserializer: D) -> Result<String, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    Ok(Option::<String>::deserialize(deserializer)?.unwrap_or_default())
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AssetBundleDetail {
-    #[serde(rename = "bundleName", deserialize_with = "de_null_as_empty_string")]
-    pub bundle_name: String,
-    #[serde(rename = "cacheFileName", deserialize_with = "de_null_as_empty_string")]
-    pub cache_file_name: String,
-    #[serde(
-        rename = "cacheDirectoryName",
-        deserialize_with = "de_null_as_empty_string"
-    )]
-    pub cache_directory_name: String,
-    // nuverse regions use `crc` instead of `hash`; the server may send nil here.
-    #[serde(rename = "hash", deserialize_with = "de_null_as_empty_string")]
-    pub hash: String,
-    #[serde(rename = "category")]
-    pub category: AssetCategory,
-    #[serde(rename = "crc")]
-    pub crc: i64,
-    #[serde(rename = "fileSize")]
-    pub file_size: i64,
-    #[serde(rename = "dependencies")]
-    pub dependencies: Vec<String>,
-    #[serde(rename = "paths", default)]
-    pub paths: Vec<String>,
-    #[serde(rename = "isBuiltin")]
-    pub is_builtin: bool,
-    #[serde(rename = "isRelocate")]
-    pub is_relocate: Option<bool>,
-    #[serde(rename = "md5Hash")]
-    pub md5_hash: Option<String>,
-    #[serde(rename = "downloadPath")]
-    pub download_path: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AssetBundleInfo {
-    #[serde(rename = "version")]
-    pub version: Option<String>,
-    #[serde(rename = "os")]
-    pub os: Option<String>,
-    #[serde(rename = "bundles")]
-    pub bundles: HashMap<String, AssetBundleDetail>,
-}
-
-#[derive(Debug, Clone)]
-struct DownloadTask {
-    download_path: String,
-    bundle_path: String,
-    bundle_hash: String,
-    category: AssetCategory,
-    file_size: i64,
-    priority: usize,
-    export_payloads: bool,
-    stage_haruki_3d: bool,
-}
-
-struct BundleWritePlan {
-    raw_target: Option<PathBuf>,
-    haruki_3d_target: Option<PathBuf>,
-    temp_target: Option<PathBuf>,
-    bundle_hash_index: Option<Arc<std::sync::Mutex<DownloadRecord>>>,
-    bundle_hash_index_key: String,
-}
-
-struct Haruki3dExportPlan {
-    config: crate::core::config::Haruki3dExportConfig,
-    info: AssetBundleInfo,
-    tasks: Vec<DownloadTask>,
-    pending_tasks: Vec<DownloadTask>,
-    pending_paths: HashSet<String>,
-    downloaded_assets: DownloadRecord,
-    record_path: PathBuf,
-    dependency_index_path: PathBuf,
-    asset_root: PathBuf,
-    work_run_dir: PathBuf,
-}
-
-#[derive(Debug, Clone)]
-pub struct AssetExecutionContext {
-    client: reqwest::Client,
-    region_name: String,
-    region: RegionConfig,
-    request: AssetUpdateRequest,
-    retry: crate::core::config::RetryConfig,
-    runtime_cookie: Option<String>,
-    resolved_asset_version: Option<String>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Haruki3dExportSummary {
-    pub matched_bundles: usize,
-    pub downloaded_bundles: usize,
-}
-
-#[derive(Debug, Clone)]
-pub enum ExecutionProgressUpdate {
-    Phase {
-        phase: JobPhase,
-        message: String,
-    },
-    DownloadsPlanned {
-        total: usize,
-    },
-    BundleStarted {
-        bundle: String,
-    },
-    BundleDownloaded {
-        bundle: String,
-        bytes: usize,
-        elapsed_ms: u128,
-    },
-    BundleTempWritten {
-        bundle: String,
-        elapsed_ms: u128,
-    },
-    BundleExported {
-        bundle: String,
-        elapsed_ms: u128,
-    },
-    BundleUnityRsExportPhases {
-        bundle: String,
-        phase_ms: HashMap<String, u64>,
-    },
-    BundleUnityRsSkippedObjectReads {
-        bundle: String,
-        count: usize,
-    },
-    BundleUnityRsObjectReadPlan {
-        bundle: String,
-        plan: NativeObjectReadPlanStats,
-    },
-    SchedulerTelemetry {
-        bundle: Option<String>,
-        phase_ms: HashMap<String, u64>,
-    },
-    BundleCompleted {
-        bundle: String,
-    },
-    BundleFailed {
-        bundle: String,
-        error: String,
-    },
-    RecordSaved {
-        entries: usize,
-    },
-    ChartHashSyncFinished {
-        performed: bool,
-    },
-}
-
-struct NativeBundlePostProcessJob {
-    bundle_path: String,
-    bundle_hash: String,
-    export_started: Instant,
-    payload_export: UnityAssetBundlePayloadExport,
-    backlog_wait_ms: u128,
-    _backlog_permit: Option<OwnedSemaphorePermit>,
-    _memory_permit: Option<OwnedSemaphorePermit>,
-}
-
-enum BundleWorkOutput {
-    Completed,
-    NativePostProcess(Box<NativeBundlePostProcessJob>),
-}
-
-impl AssetExecutionContext {
-    pub fn new(
-        app_config: &AppConfig,
-        region_name: &str,
-        region: &RegionConfig,
-        request: &AssetUpdateRequest,
-    ) -> Result<Self, AssetExecutionError> {
-        let mut headers = HeaderMap::new();
-        headers.insert(ACCEPT, HeaderValue::from_static("*/*"));
-        headers.insert(
-            USER_AGENT,
-            HeaderValue::from_static("ProductName/134 CFNetwork/1408.0.4 Darwin/22.5.0"),
-        );
-        headers.insert(ACCEPT_ENCODING, HeaderValue::from_static("identity"));
-        headers.insert(
-            ACCEPT_LANGUAGE,
-            HeaderValue::from_static("zh-CN,zh-Hans;q=0.9"),
-        );
-        headers.insert(
-            "X-Unity-Version",
-            HeaderValue::from_str(&region.runtime.unity_version)
-                .map_err(|err| AssetExecutionError::HttpClient(err.to_string()))?,
-        );
-
-        let mut builder = reqwest::Client::builder()
-            .default_headers(headers)
-            .no_gzip()
-            .no_brotli()
-            .no_deflate()
-            .no_zstd()
-            .connect_timeout(Duration::from_secs(10))
-            .timeout(Duration::from_secs(180))
-            .pool_max_idle_per_host(100)
-            .local_address(IpAddr::V4(Ipv4Addr::UNSPECIFIED))
-            .tcp_keepalive(Duration::from_secs(30));
-        if app_config.server.asset_http_version == AssetHttpVersion::Http1 {
-            builder = builder.http1_only();
-        }
-
-        if let Some(proxy) = &app_config.server.proxy {
-            if !proxy.is_empty() {
-                builder = builder.proxy(
-                    reqwest::Proxy::all(proxy)
-                        .map_err(|err| AssetExecutionError::HttpClient(err.to_string()))?,
-                );
-            }
-        }
-
-        Ok(Self {
-            client: builder
-                .build()
-                .map_err(|err| AssetExecutionError::HttpClient(err.to_string()))?,
-            region_name: region_name.to_string(),
-            region: region.clone(),
-            request: request.clone(),
-            retry: app_config.execution.retry.clone(),
-            runtime_cookie: None,
-            resolved_asset_version: request.asset_version.clone(),
-        })
-    }
-
-    pub async fn execute(
-        mut self,
-        app_config: &AppConfig,
-        progress: Option<UnboundedSender<ExecutionProgressUpdate>>,
-        cancel_flag: Option<Arc<AtomicBool>>,
-    ) -> Result<ExecutionSummary, AssetExecutionError> {
-        self.ensure_not_cancelled(&cancel_flag)?;
-        let record_path = self
-            .region
-            .paths
-            .downloaded_asset_record_file
-            .clone()
-            .ok_or_else(|| AssetExecutionError::MissingAssetSaveDir {
-                region: self.region_name.clone(),
-            })?;
-        let mut downloaded_assets = load_download_record(&record_path)?;
-        let haruki_3d_downloaded_assets = self
-            .haruki_3d_download_record_path()
-            .map(load_download_record)
-            .transpose()?
-            .unwrap_or_default();
-        let can_reuse_haruki_3d_download_record = self.can_reuse_haruki_3d_download_record().await;
-
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::FetchingAssetInfo,
-                message: "fetching asset bundle info".to_string(),
-            },
-        );
-
-        if self.requires_cookies() {
-            self.fetch_runtime_cookies().await?;
-        }
-
-        self.ensure_not_cancelled(&cancel_flag)?;
-        let info = self.fetch_asset_bundle_info().await?;
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::PlanningDownloads,
-                message: "building download task list".to_string(),
-            },
-        );
-        let tasks = self.build_download_tasks(
-            &info,
-            &downloaded_assets,
-            &haruki_3d_downloaded_assets,
-            can_reuse_haruki_3d_download_record,
-        )?;
-        let discovered_bundles = info.bundles.len();
-        let queued_downloads = tasks.len();
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::DownloadsPlanned {
-                total: queued_downloads,
-            },
-        );
-
-        if tasks.is_empty() {
-            tracing::info!(
-                region = %self.region_name,
-                discovered = discovered_bundles,
-                "no new assets to download"
-            );
-            return Ok(ExecutionSummary {
-                discovered_bundles,
-                queued_downloads: 0,
-                completed_downloads: 0,
-                failed_downloads: 0,
-                updated_record_entries: 0,
-                chart_hash_sync_performed: false,
-            });
-        }
-        // `AssetBundleInfo` can contain tens of thousands of entries. The execution phase only
-        // needs the compact task list and the discovered count, so release the larger manifest
-        // before any bundle payloads enter memory.
-        drop(info);
-
-        let mut completed = 0usize;
-        let mut completed_standard = 0usize;
-        let mut failed = 0usize;
-        let mut pending_save_count = 0usize;
-        let batch_save_size = app_config.execution.batch_save_size;
-        let concurrency = app_config.effective_concurrency();
-        let download_concurrency = concurrency.download.max(1);
-        let media_encode_concurrency = concurrency
-            .audio_encode
-            .max(concurrency.video_encode)
-            .max(concurrency.media_encode)
-            .max(1);
-        let post_process_concurrency = if concurrency.post_process == 0 {
-            media_encode_concurrency
-        } else {
-            concurrency.post_process
-        }
-        .max(1);
-        let semaphore = std::sync::Arc::new(Semaphore::new(download_concurrency));
-        let memory_limiter = BundleMemoryLimiter::from_config(app_config);
-        let post_process_semaphore = std::sync::Arc::new(Semaphore::new(post_process_concurrency));
-        let post_process_backlog_capacity =
-            post_process_backlog_capacity(download_concurrency, post_process_concurrency);
-        let post_process_backlog_semaphore =
-            std::sync::Arc::new(Semaphore::new(post_process_backlog_capacity));
-        let post_process_queued = std::sync::Arc::new(AtomicUsize::new(0));
-        let post_process_active = std::sync::Arc::new(AtomicUsize::new(0));
-        let mut joins = JoinSet::new();
-        let mut post_process_joins = JoinSet::new();
-        let export_path_registry = NativeSemanticExportPathRegistry::default();
-        // Spawned tasks require an owned `'static` config. Share one deep clone rather than cloning
-        // every region/filter/storage string once per bundle.
-        let app_config_cloned = Arc::new(app_config.clone());
-        let execution_context = Arc::new(self.clone());
-        let haruki_3d_work_root = self.haruki_3d_work_asset_root();
-        let bundle_hash_index_path = self.haruki_3d_bundle_hash_index_path();
-        let bundle_hash_index = bundle_hash_index_path
-            .as_ref()
-            .map(load_download_record)
-            .transpose()?
-            .map(|record| Arc::new(std::sync::Mutex::new(record)));
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::DownloadingBundles,
-                message: format!("downloading {queued_downloads} bundle(s)"),
-            },
-        );
-        tracing::info!(
-            region = %self.region_name,
-            queued = queued_downloads,
-            download_concurrency,
-            audio_encode_concurrency = concurrency.audio_encode,
-            video_encode_concurrency = concurrency.video_encode,
-            post_process_concurrency,
-            memory_limit_bytes = memory_limiter.limit_bytes(),
-            "starting asset bundle processing"
-        );
-
-        let spawn_bundle_task = |joins: &mut JoinSet<_>, task: DownloadTask| {
-            let ctx = execution_context.clone();
-            let semaphore = semaphore.clone();
-            let memory_limiter = memory_limiter.clone();
-            let post_process_backlog_semaphore = post_process_backlog_semaphore.clone();
-            let app_config = app_config_cloned.clone();
-            let progress = progress.clone();
-            let cancel_flag = cancel_flag.clone();
-            let haruki_3d_work_root = haruki_3d_work_root.clone();
-            let export_path_registry = export_path_registry.clone();
-            let bundle_hash_index = bundle_hash_index.clone();
-            joins.spawn(async move {
-                let download_slot_wait_started = Instant::now();
-                let _permit = semaphore.acquire_owned().await.expect("semaphore closed");
-                let download_slot_wait_ms = download_slot_wait_started.elapsed().as_millis();
-                if cancel_flag
-                    .as_ref()
-                    .is_some_and(|flag| flag.load(Ordering::SeqCst))
-                {
-                    return (
-                        task.bundle_path.clone(),
-                        task.bundle_hash.clone(),
-                        task.export_payloads,
-                        Err(AssetExecutionError::Cancelled),
-                    );
-                }
-                let memory_wait_started = Instant::now();
-                let mut memory_permit =
-                    memory_limiter.acquire(task.file_size.max(0) as usize).await;
-                let memory_wait_ms = memory_wait_started.elapsed().as_millis();
-                Self::send_progress(
-                    &progress,
-                    ExecutionProgressUpdate::BundleStarted {
-                        bundle: task.bundle_path.clone(),
-                    },
-                );
-                let bundle_path = task.bundle_path.clone();
-                let bundle_hash = task.bundle_hash.clone();
-                let record_standard = task.export_payloads;
-                let result = match ctx
-                    .download_and_export_bundle_payloads(
-                        &app_config,
-                        &task,
-                        &progress,
-                        haruki_3d_work_root.as_deref(),
-                        &export_path_registry,
-                        bundle_hash_index.as_ref(),
-                    )
-                    .await
-                {
-                    Ok(Some(mut job)) if crate::core::export_pipeline::flat_pipeline_enabled() => {
-                        // Flat mode: this worker finishes its own bundle instead of
-                        // handing it to a second pool, so there is no queue between
-                        // the stages and no backlog to bound. Parallelism is exactly
-                        // the number of workers, which is what the Python front-end
-                        // this is measured against does.
-                        job._memory_permit = memory_permit.take();
-                        match Self::finish_native_bundle_post_process(
-                            &app_config,
-                            &ctx.region_name,
-                            &ctx.region,
-                            &progress,
-                            *Box::new(job),
-                            0,
-                        )
-                        .await
-                        {
-                            Ok(()) => Ok(BundleWorkOutput::Completed),
-                            Err(error) => Err(error),
-                        }
-                    }
-                    Ok(Some(mut job)) => {
-                        let backlog_wait_started = Instant::now();
-                        let backlog_permit = post_process_backlog_semaphore
-                            .acquire_owned()
-                            .await
-                            .expect("post-process backlog semaphore closed");
-                        let backlog_wait_ms = backlog_wait_started.elapsed().as_millis();
-                        job.backlog_wait_ms = backlog_wait_ms;
-                        job._backlog_permit = Some(backlog_permit);
-                        job._memory_permit = memory_permit.take();
-                        Ok(BundleWorkOutput::NativePostProcess(Box::new(job)))
-                    }
-                    Ok(None) => Ok(BundleWorkOutput::Completed),
-                    Err(error) => Err(error),
-                };
-                let mut phase_ms = HashMap::new();
-                phase_ms.insert(
-                    "scheduler.download_slot_wait".to_string(),
-                    download_slot_wait_ms.min(u128::from(u64::MAX)) as u64,
-                );
-                phase_ms.insert(
-                    "scheduler.memory_wait".to_string(),
-                    memory_wait_ms.min(u128::from(u64::MAX)) as u64,
-                );
-                Self::send_progress(
-                    &progress,
-                    ExecutionProgressUpdate::SchedulerTelemetry {
-                        bundle: Some(task.bundle_path.clone()),
-                        phase_ms,
-                    },
-                );
-                (bundle_path, bundle_hash, record_standard, result)
-            });
-        };
-
-        // Maintain a fixed refill window instead of creating one Tokio task per bundle up front.
-        // The semaphore remains as a defensive invariant, while pending task metadata stays O(D)
-        // where D is download concurrency rather than O(all discovered bundles).
-        let mut remaining_tasks = tasks.into_iter();
-        for task in remaining_tasks.by_ref().take(download_concurrency) {
-            spawn_bundle_task(&mut joins, task);
-        }
-
-        while !joins.is_empty() || !post_process_joins.is_empty() {
-            // If cancellation was requested, stop scheduling/awaiting more (expensive) post-process
-            // work and fall through to persist the record before returning Cancelled, rather than
-            // draining every already-queued bundle first.
-            if self.ensure_not_cancelled(&cancel_flag).is_err() {
-                break;
-            }
-            tokio::select! {
-                Some(result) = joins.join_next(), if !joins.is_empty() => {
-                    let completed_task = match result {
-                        Ok(tuple) => Some(tuple),
-                        Err(join_err) => {
-                            // A download/export sub-task panicked or was aborted. Count it as a
-                            // failed bundle instead of unwinding the orchestrator (which would
-                            // leave the owning job wedged in Running forever).
-                            failed += 1;
-                            tracing::error!(
-                                region = %self.region_name,
-                                error = %join_err,
-                                "bundle download/export task panicked or was aborted; counting as failed"
-                            );
-                            None
-                        }
-                    };
-                    if let Some((bundle_path, bundle_hash, record_standard, result)) = completed_task {
-                      match result {
-                       Ok(BundleWorkOutput::Completed) => {
-                            Self::record_completed_bundle(
-                                &progress,
-                                &record_path,
-                                &mut downloaded_assets,
-                                &mut completed,
-                                &mut completed_standard,
-                                &mut pending_save_count,
-                                batch_save_size,
-                                &self.region_name,
-                                bundle_hash_index_path.as_ref(),
-                                bundle_hash_index.as_ref(),
-                                bundle_path,
-                                bundle_hash,
-                                record_standard,
-                            )
-                            .await;
-                        }
-                       Ok(BundleWorkOutput::NativePostProcess(job)) => {
-                            let app_config = app_config_cloned.clone();
-                            let region = self.region.clone();
-                            let region_name = self.region_name.clone();
-                            let progress = progress.clone();
-                            let semaphore = post_process_semaphore.clone();
-                            let post_process_queued = post_process_queued.clone();
-                            let post_process_active = post_process_active.clone();
-                            let queued = post_process_queued.fetch_add(1, Ordering::Relaxed) + 1;
-                            Self::send_progress(
-                                &progress,
-                                ExecutionProgressUpdate::SchedulerTelemetry {
-                                    bundle: Some(job.bundle_path.clone()),
-                                    phase_ms: HashMap::from([
-                                        (
-                                            "scheduler.post_process_queued".to_string(),
-                                            queued as u64,
-                                        ),
-                                        (
-                                            "scheduler.post_process_backlog_capacity".to_string(),
-                                            post_process_backlog_capacity as u64,
-                                        ),
-                                        (
-                                            "scheduler.post_process_concurrency".to_string(),
-                                            post_process_concurrency as u64,
-                                        ),
-                                    ]),
-                                },
-                            );
-                            post_process_joins.spawn(async move {
-                                let queue_started = Instant::now();
-                                let _permit = semaphore.acquire_owned().await.expect("semaphore closed");
-                                let queue_wait_ms = queue_started.elapsed().as_millis();
-                                post_process_queued.fetch_sub(1, Ordering::Relaxed);
-                                let active = post_process_active.fetch_add(1, Ordering::Relaxed) + 1;
-                                let bundle_path = job.bundle_path.clone();
-                                let bundle_hash = job.bundle_hash.clone();
-                                let result = Self::finish_native_bundle_post_process(
-                                    &app_config,
-                                    &region_name,
-                                    &region,
-                                    &progress,
-                                    *job,
-                                    queue_wait_ms,
-                                )
-                                .await;
-                                post_process_active.fetch_sub(1, Ordering::Relaxed);
-                                Self::send_progress(
-                                    &progress,
-                                    ExecutionProgressUpdate::SchedulerTelemetry {
-                                        bundle: Some(bundle_path.clone()),
-                                        phase_ms: HashMap::from([
-                                            (
-                                                "scheduler.post_process_active_peak".to_string(),
-                                                active as u64,
-                                            ),
-                                            (
-                                                "scheduler.post_process_queue_wait".to_string(),
-                                                queue_wait_ms.min(u128::from(u64::MAX)) as u64,
-                                            ),
-                                        ]),
-                                    },
-                                );
-                                (bundle_path, bundle_hash, true, result)
-                            });
-                        }
-                       Err(AssetExecutionError::Cancelled) => {
-                            // Stop scheduling further work but fall through to persist the record so
-                            // already-completed bundles aren't re-downloaded on the next run.
-                            break;
-                        }
-                       Err(err) => {
-                            failed += 1;
-                            Self::send_progress(
-                                &progress,
-                                ExecutionProgressUpdate::BundleFailed {
-                                    bundle: bundle_path.clone(),
-                                    error: err.to_string(),
-                                },
-                            );
-                            tracing::warn!(
-                                region = %self.region_name,
-                                bundle = %bundle_path,
-                                error = %err,
-                                "bundle processing failed"
-                            );
-                        }
-                      }
-                    }
-                    if let Some(task) = remaining_tasks.next() {
-                        spawn_bundle_task(&mut joins, task);
-                    }
-                }
-                Some(result) = post_process_joins.join_next(), if !post_process_joins.is_empty() => {
-                    let (bundle_path, bundle_hash, record_standard, result) = match result {
-                        Ok(tuple) => tuple,
-                        Err(join_err) => {
-                            // Post-process sub-task panicked or was aborted: count as failed
-                            // rather than re-panicking the orchestrator.
-                            failed += 1;
-                            tracing::error!(
-                                region = %self.region_name,
-                                error = %join_err,
-                                "bundle post-process task panicked or was aborted; counting as failed"
-                            );
-                            continue;
-                        }
-                    };
-                    match result {
-                        Ok(()) => {
-                            Self::record_completed_bundle(
-                                &progress,
-                                &record_path,
-                                &mut downloaded_assets,
-                                &mut completed,
-                                &mut completed_standard,
-                                &mut pending_save_count,
-                                batch_save_size,
-                                &self.region_name,
-                                bundle_hash_index_path.as_ref(),
-                                bundle_hash_index.as_ref(),
-                                bundle_path,
-                                bundle_hash,
-                                record_standard,
-                            )
-                            .await;
-                        }
-                        Err(AssetExecutionError::Cancelled) => {
-                            // Stop scheduling further work but fall through to persist the record so
-                            // already-completed bundles aren't re-downloaded on the next run.
-                            break;
-                        }
-                        Err(err) => {
-                            failed += 1;
-                            Self::send_progress(
-                                &progress,
-                                ExecutionProgressUpdate::BundleFailed {
-                                    bundle: bundle_path.clone(),
-                                    error: err.to_string(),
-                                },
-                            );
-                            tracing::warn!(
-                                region = %self.region_name,
-                                bundle = %bundle_path,
-                                error = %err,
-                                "bundle post-process failed"
-                            );
-                        }
-                    }
-                }
-            }
-        }
-
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::PersistingState,
-                message: "saving downloaded asset record".to_string(),
-            },
-        );
-        // Persist the record BEFORE honoring cancellation: every bundle that finished already ran
-        // its export/upload side effects, so dropping them here would force a redundant re-run.
-        Self::save_bundle_hash_index_checkpoint(
-            bundle_hash_index_path.as_ref(),
-            bundle_hash_index.as_ref(),
-        )
-        .await?;
-        Self::save_download_record_on_blocking_thread(record_path.clone(), &mut downloaded_assets)
-            .await?;
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::RecordSaved {
-                entries: downloaded_assets.len(),
-            },
-        );
-        // Honor cancellation now that the record is durable (this skips the heavier chart sync).
-        self.ensure_not_cancelled(&cancel_flag)?;
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::SyncingChartHashes,
-                message: "syncing chart hashes".to_string(),
-            },
-        );
-        let chart_hash_sync_performed = sync_chart_hashes(
-            &app_config.git_sync.chart_hashes,
-            &self.region_name,
-            &downloaded_assets,
-            app_config.server.proxy.as_deref(),
-            &app_config.execution.retry,
-            false,
-        )?
-        .is_some();
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::ChartHashSyncFinished {
-                performed: chart_hash_sync_performed,
-            },
-        );
-
-        Ok(ExecutionSummary {
-            discovered_bundles,
-            queued_downloads,
-            completed_downloads: completed,
-            failed_downloads: failed,
-            // Number of record entries actually added/updated this run (not the whole record size),
-            // keeping the semantics consistent with the empty-task early-return path above.
-            updated_record_entries: completed_standard,
-            chart_hash_sync_performed,
-        })
-    }
-
-    pub async fn prefetch_asset_bundles(
-        mut self,
-        app_config: &AppConfig,
-        progress: Option<UnboundedSender<ExecutionProgressUpdate>>,
-        cancel_flag: Option<Arc<AtomicBool>>,
-    ) -> Result<ExecutionSummary, AssetExecutionError> {
-        self.ensure_not_cancelled(&cancel_flag)?;
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::FetchingAssetInfo,
-                message: "fetching asset bundle info".to_string(),
-            },
-        );
-
-        if self.requires_cookies() {
-            self.fetch_runtime_cookies().await?;
-        }
-
-        self.ensure_not_cancelled(&cancel_flag)?;
-        let info = self.fetch_asset_bundle_info().await?;
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::PlanningDownloads,
-                message: "building prefetch task list".to_string(),
-            },
-        );
-        let tasks = if configured_asset_bundle_cache_dir(app_config).is_some() {
-            self.build_standard_download_tasks(&info, &DownloadRecord::new())
-        } else {
-            self.build_raw_bundle_filter_tasks(&info)
-        };
-        let discovered_bundles = info.bundles.len();
-        let queued_downloads = tasks.len();
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::DownloadsPlanned {
-                total: queued_downloads,
-            },
-        );
-
-        if tasks.is_empty() {
-            tracing::info!(
-                region = %self.region_name,
-                discovered = discovered_bundles,
-                "no assets matched prefetch filters"
-            );
-            return Ok(ExecutionSummary {
-                discovered_bundles,
-                queued_downloads: 0,
-                completed_downloads: 0,
-                failed_downloads: 0,
-                updated_record_entries: 0,
-                chart_hash_sync_performed: false,
-            });
-        }
-        drop(info);
-
-        let download_concurrency = app_config.effective_concurrency().download.max(1);
-        let semaphore = Arc::new(Semaphore::new(download_concurrency));
-        let memory_limiter = BundleMemoryLimiter::from_config(app_config);
-        let mut joins = JoinSet::new();
-        let app_config_cloned = Arc::new(app_config.clone());
-        let execution_context = Arc::new(self.clone());
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::DownloadingBundles,
-                message: format!("prefetching {queued_downloads} bundle(s)"),
-            },
-        );
-        tracing::info!(
-            region = %self.region_name,
-            queued = queued_downloads,
-            download_concurrency,
-            memory_limit_bytes = memory_limiter.limit_bytes(),
-            "starting asset bundle prefetch"
-        );
-
-        let spawn_prefetch_task = |joins: &mut JoinSet<_>, task: DownloadTask| {
-            let ctx = execution_context.clone();
-            let semaphore = semaphore.clone();
-            let memory_limiter = memory_limiter.clone();
-            let app_config = app_config_cloned.clone();
-            let progress = progress.clone();
-            let cancel_flag = cancel_flag.clone();
-            joins.spawn(async move {
-                let _permit = semaphore.acquire_owned().await.expect("semaphore closed");
-                if cancel_flag
-                    .as_ref()
-                    .is_some_and(|flag| flag.load(Ordering::SeqCst))
-                {
-                    return (
-                        task.bundle_path.clone(),
-                        Err(AssetExecutionError::Cancelled),
-                    );
-                }
-                let _memory_permit = memory_limiter.acquire(task.file_size.max(0) as usize).await;
-                Self::send_progress(
-                    &progress,
-                    ExecutionProgressUpdate::BundleStarted {
-                        bundle: task.bundle_path.clone(),
-                    },
-                );
-                let bundle_path = task.bundle_path.clone();
-                let result = ctx.prefetch_bundle(&app_config, &task, &progress).await;
-                (bundle_path, result)
-            });
-        };
-        let mut remaining_tasks = tasks.into_iter();
-        for task in remaining_tasks.by_ref().take(download_concurrency) {
-            spawn_prefetch_task(&mut joins, task);
-        }
-
-        let mut completed = 0usize;
-        let mut failed = 0usize;
-        while let Some(result) = joins.join_next().await {
-            let completed_task = match result {
-                Ok(tuple) => Some(tuple),
-                Err(join_err) => {
-                    // Prefetch sub-task panicked or was aborted: count as failed instead of
-                    // unwinding the run.
-                    failed += 1;
-                    tracing::error!(
-                        region = %self.region_name,
-                        error = %join_err,
-                        "bundle prefetch task panicked or was aborted; counting as failed"
-                    );
-                    None
-                }
-            };
-            if let Some((bundle_path, result)) = completed_task {
-                match result {
-                    Ok(()) => {
-                        completed += 1;
-                        Self::send_progress(
-                            &progress,
-                            ExecutionProgressUpdate::BundleCompleted {
-                                bundle: bundle_path,
-                            },
-                        );
-                    }
-                    Err(AssetExecutionError::Cancelled) => {
-                        return Err(AssetExecutionError::Cancelled);
-                    }
-                    Err(err) => {
-                        failed += 1;
-                        Self::send_progress(
-                            &progress,
-                            ExecutionProgressUpdate::BundleFailed {
-                                bundle: bundle_path.clone(),
-                                error: err.to_string(),
-                            },
-                        );
-                        tracing::warn!(
-                            region = %self.region_name,
-                            bundle = %bundle_path,
-                            error = %err,
-                            "bundle prefetch failed"
-                        );
-                    }
-                }
-            }
-            if let Some(task) = remaining_tasks.next() {
-                spawn_prefetch_task(&mut joins, task);
-            }
-        }
-
-        Ok(ExecutionSummary {
-            discovered_bundles,
-            queued_downloads,
-            completed_downloads: completed,
-            failed_downloads: failed,
-            updated_record_entries: 0,
-            chart_hash_sync_performed: false,
-        })
-    }
-
-    fn ensure_not_cancelled(
-        &self,
-        cancel_flag: &Option<Arc<AtomicBool>>,
-    ) -> Result<(), AssetExecutionError> {
-        if cancel_flag
-            .as_ref()
-            .is_some_and(|flag| flag.load(Ordering::SeqCst))
-        {
-            Err(AssetExecutionError::Cancelled)
-        } else {
-            Ok(())
-        }
-    }
-
-    fn send_progress(
-        sender: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-        update: ExecutionProgressUpdate,
-    ) {
-        if let Some(sender) = sender {
-            let _ = sender.send(update);
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn record_completed_bundle(
-        progress: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-        record_path: &str,
-        downloaded_assets: &mut DownloadRecord,
-        completed: &mut usize,
-        completed_standard: &mut usize,
-        pending_save_count: &mut usize,
-        batch_save_size: usize,
-        region_name: &str,
-        bundle_hash_index_path: Option<&PathBuf>,
-        bundle_hash_index: Option<&Arc<std::sync::Mutex<DownloadRecord>>>,
-        bundle_path: String,
-        bundle_hash: String,
-        record_standard: bool,
-    ) {
-        *completed += 1;
-        if record_standard {
-            *completed_standard += 1;
-            downloaded_assets.insert(bundle_path.clone(), bundle_hash);
-        }
-        *pending_save_count += 1;
-        if progress.is_none() {
-            tracing::info!(
-                region = %region_name,
-                bundle = %bundle_path,
-                completed = *completed,
-                "bundle completed"
-            );
-        }
-        Self::send_progress(
-            progress,
-            ExecutionProgressUpdate::BundleCompleted {
-                bundle: bundle_path,
-            },
-        );
-        if batch_save_size > 0 && *pending_save_count >= batch_save_size {
-            tracing::info!(
-                region = %region_name,
-                batch = *pending_save_count,
-                "batch-flushing download record"
-            );
-            let save_result = match Self::save_bundle_hash_index_checkpoint(
-                bundle_hash_index_path,
-                bundle_hash_index,
-            )
-            .await
-            {
-                Ok(()) => {
-                    Self::save_download_record_on_blocking_thread(
-                        record_path.to_string(),
-                        downloaded_assets,
-                    )
-                    .await
-                }
-                Err(error) => Err(error),
-            };
-            match save_result {
-                Ok(()) => Self::send_progress(
-                    progress,
-                    ExecutionProgressUpdate::RecordSaved {
-                        entries: downloaded_assets.len(),
-                    },
-                ),
-                Err(err) => tracing::warn!(
-                    region = %region_name,
-                    error = %err,
-                    "mid-run batch save of download record failed; will retry at end"
-                ),
-            }
-            *pending_save_count = 0;
-        }
-    }
-
-    async fn save_download_record_on_blocking_thread(
-        record_path: String,
-        downloaded_assets: &mut DownloadRecord,
-    ) -> Result<(), AssetExecutionError> {
-        let record = std::mem::take(downloaded_assets);
-        let (record, result) = tokio::task::spawn_blocking(move || {
-            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                save_download_record(record_path, &record)
-            }));
-            (record, result)
-        })
-        .await
-        .map_err(|source| AssetExecutionError::BlockingTask(source.to_string()))?;
-        *downloaded_assets = record;
-        result
-            .map_err(|panic| AssetExecutionError::BlockingTask(blocking_panic_message(panic)))?
-            .map_err(AssetExecutionError::from)
-    }
-
-    async fn save_owned_download_record_on_blocking_thread(
-        record_path: String,
-        downloaded_assets: DownloadRecord,
-    ) -> Result<(), AssetExecutionError> {
-        tokio::task::spawn_blocking(move || save_download_record(record_path, &downloaded_assets))
-            .await
-            .map_err(|source| AssetExecutionError::BlockingTask(source.to_string()))?
-            .map_err(AssetExecutionError::from)
-    }
-
-    async fn save_bundle_hash_index_checkpoint(
-        path: Option<&PathBuf>,
-        index: Option<&Arc<std::sync::Mutex<DownloadRecord>>>,
-    ) -> Result<(), AssetExecutionError> {
-        let (Some(path), Some(index)) = (path, index) else {
-            return Ok(());
-        };
-        let record = index
-            .lock()
-            .map_err(|_| {
-                AssetExecutionError::BlockingTask("bundle hash index lock poisoned".to_string())
-            })?
-            .clone();
-        Self::save_owned_download_record_on_blocking_thread(
-            path.to_string_lossy().into_owned(),
-            record,
-        )
-        .await
-    }
-
-    async fn finish_native_bundle_post_process(
-        app_config: &AppConfig,
-        region_name: &str,
-        region: &RegionConfig,
-        progress: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-        job: NativeBundlePostProcessJob,
-        queue_wait_ms: u128,
-    ) -> Result<(), AssetExecutionError> {
-        let post_process_summary = post_process_exported_files(
-            app_config,
-            region_name,
-            region,
-            &job.payload_export.export_path,
-            &job.payload_export.export_root,
-            job.payload_export.native_scoped_post_process,
-            &job.payload_export.native_written_files,
-            job.payload_export.native_acb_sources,
-        )
-        .await?;
-
-        let mut phase_ms = job.payload_export.unity_rs_export_phase_ms;
-        phase_ms.extend(post_process_summary.post_process_phase_ms);
-        phase_ms.insert(
-            "post_process.queue_wait".to_string(),
-            queue_wait_ms.min(u128::from(u64::MAX)) as u64,
-        );
-        phase_ms.insert(
-            "scheduler.post_process_backlog_wait".to_string(),
-            job.backlog_wait_ms.min(u128::from(u64::MAX)) as u64,
-        );
-        phase_ms.insert(
-            "scheduler.bundle_active_before_post_process".to_string(),
-            job.export_started
-                .elapsed()
-                .as_millis()
-                .min(u128::from(u64::MAX)) as u64,
-        );
-        if !phase_ms.is_empty() {
-            Self::send_progress(
-                progress,
-                ExecutionProgressUpdate::BundleUnityRsExportPhases {
-                    bundle: job.bundle_path.clone(),
-                    phase_ms,
-                },
-            );
-        }
-        if !job.payload_export.unity_rs_skipped_object_reads.is_empty() {
-            Self::send_progress(
-                progress,
-                ExecutionProgressUpdate::BundleUnityRsSkippedObjectReads {
-                    bundle: job.bundle_path.clone(),
-                    count: job.payload_export.unity_rs_skipped_object_reads.len(),
-                },
-            );
-        }
-        if !job.payload_export.unity_rs_object_read_plan.is_empty() {
-            Self::send_progress(
-                progress,
-                ExecutionProgressUpdate::BundleUnityRsObjectReadPlan {
-                    bundle: job.bundle_path.clone(),
-                    plan: job.payload_export.unity_rs_object_read_plan,
-                },
-            );
-        }
-        Self::send_progress(
-            progress,
-            ExecutionProgressUpdate::BundleExported {
-                bundle: job.bundle_path,
-                elapsed_ms: job.export_started.elapsed().as_millis(),
-            },
-        );
-        Ok(())
-    }
-
-    fn requires_cookies(&self) -> bool {
-        match &self.region.provider {
-            RegionProviderConfig::ColorfulPalette {
-                required_cookies, ..
-            } => *required_cookies,
-            RegionProviderConfig::Nuverse {
-                required_cookies, ..
-            } => *required_cookies,
-        }
-    }
-
-    async fn fetch_runtime_cookies(&mut self) -> Result<(), AssetExecutionError> {
-        let url = match &self.region.provider {
-            RegionProviderConfig::ColorfulPalette {
-                cookie_bootstrap_url,
-                ..
-            }
-            | RegionProviderConfig::Nuverse {
-                cookie_bootstrap_url,
-                ..
-            } => cookie_bootstrap_url.clone().unwrap_or_else(|| {
-                "https://issue.sekai.colorfulpalette.org/api/signature".to_string()
-            }),
-        };
-        self.runtime_cookie = retry_async(
-            &self.retry,
-            "cookie bootstrap",
-            |_| async {
-                let response = self.client.post(&url).send().await.map_err(|err| {
-                    tracing::warn!(
-                        url,
-                        error = %format_reqwest_error_chain(&err),
-                        "HTTP request failed"
-                    );
-                    AssetExecutionError::Http(err)
-                })?;
-                if response.status().is_success() {
-                    Ok(response
-                        .headers()
-                        .get(SET_COOKIE)
-                        .and_then(|value| value.to_str().ok())
-                        .map(str::to_string))
-                } else {
-                    Err(AssetExecutionError::HttpStatus {
-                        url: url.clone(),
-                        status: response.status().as_u16(),
-                    })
-                }
-            },
-            is_retryable_http_error,
-        )
-        .await?;
-        Ok(())
-    }
-
-    async fn fetch_asset_bundle_info(&mut self) -> Result<AssetBundleInfo, AssetExecutionError> {
-        let url = self.render_asset_info_url().await?;
-        let body = self.get_with_retry(&url).await?;
-        decrypt_asset_bundle_info(
-            self.region.crypto.aes_key_hex.as_deref().ok_or_else(|| {
-                AssetExecutionError::MissingCryptoConfig {
-                    region: self.region_name.clone(),
-                }
-            })?,
-            self.region.crypto.aes_iv_hex.as_deref().ok_or_else(|| {
-                AssetExecutionError::MissingCryptoConfig {
-                    region: self.region_name.clone(),
-                }
-            })?,
-            &body,
-        )
-    }
-
-    async fn render_asset_info_url(&mut self) -> Result<String, AssetExecutionError> {
-        match &self.region.provider {
-            RegionProviderConfig::ColorfulPalette {
-                asset_info_url_template,
-                profile,
-                profile_hashes,
-                ..
-            } => {
-                let asset_version = self.request.asset_version.as_deref().ok_or_else(|| {
-                    AssetExecutionError::MissingAssetVersionOrHash {
-                        region: self.region_name.clone(),
-                    }
-                })?;
-                let asset_hash = self.request.asset_hash.as_deref().ok_or_else(|| {
-                    AssetExecutionError::MissingAssetVersionOrHash {
-                        region: self.region_name.clone(),
-                    }
-                })?;
-                let profile_hash = profile_hashes.get(profile).ok_or_else(|| {
-                    AssetExecutionError::MissingProfileHash {
-                        region: self.region_name.clone(),
-                        profile: profile.clone(),
-                    }
-                })?;
-                Ok(asset_info_url_template
-                    .replace("{env}", profile)
-                    .replace("{hash}", profile_hash)
-                    .replace("{asset_version}", asset_version)
-                    .replace("{asset_hash}", asset_hash)
-                    + &time_arg_jst())
-            }
-            RegionProviderConfig::Nuverse {
-                asset_version_url,
-                app_version,
-                asset_info_url_template,
-                ..
-            } => {
-                // For nuverse, always fetch the version from asset_version_url.
-                // The incoming request.asset_version is intentionally ignored here
-                // to match Go reference behavior.
-                let version_url = asset_version_url.replace("{app_version}", app_version);
-                let resolved_version =
-                    String::from_utf8_lossy(&self.get_with_retry(&version_url).await?)
-                        .trim()
-                        .to_string();
-                self.resolved_asset_version = Some(resolved_version.clone());
-                Ok(asset_info_url_template
-                    .replace("{app_version}", app_version)
-                    .replace("{asset_version}", &resolved_version)
-                    + &time_arg_jst())
-            }
-        }
-    }
-
-    fn render_bundle_url(&self, task: &DownloadTask) -> Result<String, AssetExecutionError> {
-        match &self.region.provider {
-            RegionProviderConfig::ColorfulPalette {
-                asset_bundle_url_template,
-                profile,
-                profile_hashes,
-                ..
-            } => {
-                let asset_version = self.request.asset_version.as_deref().ok_or_else(|| {
-                    AssetExecutionError::MissingAssetVersionOrHash {
-                        region: self.region_name.clone(),
-                    }
-                })?;
-                let asset_hash = self.request.asset_hash.as_deref().ok_or_else(|| {
-                    AssetExecutionError::MissingAssetVersionOrHash {
-                        region: self.region_name.clone(),
-                    }
-                })?;
-                let profile_hash = profile_hashes.get(profile).ok_or_else(|| {
-                    AssetExecutionError::MissingProfileHash {
-                        region: self.region_name.clone(),
-                        profile: profile.clone(),
-                    }
-                })?;
-
-                Ok(asset_bundle_url_template
-                    .replace("{bundle_path}", &task.download_path)
-                    .replace("{asset_version}", asset_version)
-                    .replace("{asset_hash}", asset_hash)
-                    .replace("{env}", profile)
-                    .replace("{hash}", profile_hash)
-                    + &time_arg_jst())
-            }
-            RegionProviderConfig::Nuverse {
-                asset_bundle_url_template,
-                app_version,
-                ..
-            } => {
-                let asset_version = self
-                    .resolved_asset_version
-                    .as_deref()
-                    .unwrap_or("<resolved-asset-version>");
-                Ok(asset_bundle_url_template
-                    .replace("{bundle_path}", &task.download_path)
-                    .replace("{app_version}", app_version)
-                    .replace("{asset_version}", asset_version)
-                    + &time_arg_jst())
-            }
-        }
-    }
-
-    async fn get_with_retry(&self, url: &str) -> Result<Vec<u8>, AssetExecutionError> {
-        retry_async(
-            &self.retry,
-            "http get",
-            |_| async {
-                let mut request = self.client.get(url);
-                if let Some(cookie) = &self.runtime_cookie {
-                    request = request.header(COOKIE, cookie);
-                }
-                match request.send().await {
-                    Ok(mut response) if response.status().is_success() => {
-                        // Build the final owned buffer directly from response chunks. Calling
-                        // `bytes().to_vec()` first aggregates into `Bytes` and then copies the
-                        // entire response a second time.
-                        const MAX_HTTP_PREALLOC_BYTES: u64 = 64 * 1024 * 1024;
-                        let capacity = response
-                            .content_length()
-                            .map(|length| length.min(MAX_HTTP_PREALLOC_BYTES))
-                            .and_then(|length| usize::try_from(length).ok())
-                            .unwrap_or_default();
-                        let mut body = Vec::with_capacity(capacity);
-                        while let Some(chunk) = response.chunk().await? {
-                            body.extend_from_slice(&chunk);
-                        }
-                        Ok(body)
-                    }
-                    Ok(response) => Err(AssetExecutionError::HttpStatus {
-                        url: url.to_string(),
-                        status: response.status().as_u16(),
-                    }),
-                    Err(err) => {
-                        tracing::warn!(
-                            url,
-                            error = %format_reqwest_error_chain(&err),
-                            "HTTP request failed"
-                        );
-                        Err(AssetExecutionError::Http(err))
-                    }
-                }
-            },
-            is_retryable_http_error,
-        )
-        .await
-    }
-
-    fn build_download_tasks(
-        &self,
-        info: &AssetBundleInfo,
-        downloaded_assets: &DownloadRecord,
-        haruki_3d_downloaded_assets: &DownloadRecord,
-        can_reuse_haruki_3d_download_record: bool,
-    ) -> Result<Vec<DownloadTask>, AssetExecutionError> {
-        self.append_haruki_3d_download_tasks(
-            self.build_standard_download_tasks(info, downloaded_assets),
-            info,
-            haruki_3d_downloaded_assets,
-            can_reuse_haruki_3d_download_record,
-        )
-    }
-
-    fn build_standard_download_tasks(
-        &self,
-        info: &AssetBundleInfo,
-        downloaded_assets: &DownloadRecord,
-    ) -> Vec<DownloadTask> {
-        let skip_patterns = compile_patterns(&self.region.filters.skip);
-        let priority_patterns = compile_patterns(&self.region.filters.priority);
-        let start_app_patterns = compile_patterns(&self.region.filters.start_app);
-        let on_demand_patterns = compile_patterns(&self.region.filters.on_demand);
-        let mut tasks = Vec::new();
-
-        for (bundle_name, detail) in &info.bundles {
-            if matches_any(&skip_patterns, bundle_name) {
-                continue;
-            }
-            let category_patterns = match &detail.category {
-                AssetCategory::StartApp => &start_app_patterns,
-                AssetCategory::OnDemand | AssetCategory::LivePv => &on_demand_patterns,
-                AssetCategory::Other(_) => continue,
-            };
-            if category_patterns.is_empty() || !matches_any(category_patterns, bundle_name) {
-                continue;
-            }
-
-            let bundle_hash = match self.region.provider {
-                RegionProviderConfig::Nuverse { .. } => detail.crc.to_string(),
-                RegionProviderConfig::ColorfulPalette { .. } => detail.hash.clone(),
-            };
-
-            if downloaded_assets
-                .get(bundle_name)
-                .is_some_and(|existing| existing == &bundle_hash)
-            {
-                continue;
-            }
-
-            let priority = first_match_index(&priority_patterns, bundle_name).unwrap_or(usize::MAX);
-            tasks.push(DownloadTask {
-                download_path: download_path_for_region(&self.region.provider, bundle_name, detail),
-                bundle_path: bundle_name.clone(),
-                bundle_hash,
-                category: detail.category.clone(),
-                file_size: detail.file_size,
-                priority,
-                export_payloads: true,
-                stage_haruki_3d: false,
-            });
-        }
-
-        tasks.sort_by(|a, b| {
-            a.priority
-                .cmp(&b.priority)
-                .then_with(|| a.bundle_path.cmp(&b.bundle_path))
-        });
-        tasks
-    }
-
-    fn append_haruki_3d_download_tasks(
-        &self,
-        mut tasks: Vec<DownloadTask>,
-        info: &AssetBundleInfo,
-        downloaded_assets: &DownloadRecord,
-        can_reuse_download_record: bool,
-    ) -> Result<Vec<DownloadTask>, AssetExecutionError> {
-        tasks.extend(self.build_haruki_3d_download_tasks(
-            info,
-            downloaded_assets,
-            can_reuse_download_record,
-        )?);
-        tasks.sort_by(|a, b| {
-            a.bundle_path
-                .cmp(&b.bundle_path)
-                .then_with(|| a.priority.cmp(&b.priority))
-        });
-        tasks.dedup_by(|a, b| {
-            if a.bundle_path != b.bundle_path {
-                return false;
-            }
-            b.export_payloads |= a.export_payloads;
-            b.stage_haruki_3d |= a.stage_haruki_3d;
-            true
-        });
-        tasks.sort_by(|a, b| {
-            a.priority
-                .cmp(&b.priority)
-                .then_with(|| a.bundle_path.cmp(&b.bundle_path))
-        });
-        Ok(tasks)
-    }
-
-    async fn download_and_export_bundle_payloads(
-        &self,
-        app_config: &AppConfig,
-        task: &DownloadTask,
-        progress: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-        haruki_3d_work_root: Option<&Path>,
-        export_path_registry: &NativeSemanticExportPathRegistry,
-        bundle_hash_index: Option<&Arc<std::sync::Mutex<DownloadRecord>>>,
-    ) -> Result<Option<NativeBundlePostProcessJob>, AssetExecutionError> {
-        let asset_save_dir = self.region.paths.asset_save_dir.clone().ok_or_else(|| {
-            AssetExecutionError::MissingAssetSaveDir {
-                region: self.region_name.clone(),
-            }
-        })?;
-        let bundle_url = self.render_bundle_url(task)?;
-        let download_started = Instant::now();
-        let fetch = self
-            .fetch_deobfuscated_bundle(app_config, &bundle_url, task)
-            .await?;
-        Self::send_progress(
-            progress,
-            ExecutionProgressUpdate::BundleDownloaded {
-                bundle: task.bundle_path.clone(),
-                bytes: fetch.body.len(),
-                elapsed_ms: download_started.elapsed().as_millis(),
-            },
-        );
-
-        let temp_file = self.bundle_temp_file(task)?;
-        let write_plan = self.bundle_write_plan(
-            task,
-            &asset_save_dir,
-            haruki_3d_work_root,
-            &temp_file,
-            bundle_hash_index,
-        )?;
-        let blocking_started = Instant::now();
-        Self::persist_bundle_payload(fetch.body, write_plan).await?;
-        if task.export_payloads {
-            Self::send_progress(
-                progress,
-                ExecutionProgressUpdate::BundleTempWritten {
-                    bundle: task.bundle_path.clone(),
-                    elapsed_ms: blocking_started.elapsed().as_millis(),
-                },
-            );
-        }
-
-        if !task.export_payloads {
-            return Ok(None);
-        }
-
-        self.export_bundle_payloads(
-            app_config,
-            task,
-            &asset_save_dir,
-            &temp_file,
-            export_path_registry,
-        )
-        .await
-    }
-
-    fn bundle_temp_file(&self, task: &DownloadTask) -> Result<PathBuf, AssetExecutionError> {
-        // Asset-info is untrusted, so validate before using the bundle name in any path.
-        let safe_path = validate_relative_bundle_path(&task.bundle_path)?;
-        Ok(std::env::temp_dir().join(&self.region_name).join(safe_path))
-    }
-
-    fn bundle_write_plan(
-        &self,
-        task: &DownloadTask,
-        asset_save_dir: &str,
-        haruki_3d_work_root: Option<&Path>,
-        temp_file: &Path,
-        bundle_hash_index: Option<&Arc<std::sync::Mutex<DownloadRecord>>>,
-    ) -> Result<BundleWritePlan, AssetExecutionError> {
-        let raw_target = self
-            .matches_raw_bundle_filters(&task.bundle_path)
-            .then(|| self.raw_bundle_output_path(asset_save_dir, &task.bundle_path))
-            .transpose()?;
-        let haruki_3d_target = task
-            .stage_haruki_3d
-            .then_some(haruki_3d_work_root)
-            .flatten()
-            .map(|root| raw_bundle_output_path(root, &task.bundle_path))
-            .transpose()?;
-        Ok(BundleWritePlan {
-            raw_target,
-            haruki_3d_target,
-            temp_target: task.export_payloads.then(|| temp_file.to_path_buf()),
-            bundle_hash_index: bundle_hash_index.cloned(),
-            bundle_hash_index_key: bundle_hash_index_key(&task.bundle_path)?,
-        })
-    }
-
-    async fn persist_bundle_payload(
-        payload: Vec<u8>,
-        plan: BundleWritePlan,
-    ) -> Result<(), AssetExecutionError> {
-        tokio::task::spawn_blocking(move || Self::write_bundle_payload(&payload, plan))
-            .await
-            .map_err(|source| AssetExecutionError::BlockingTask(source.to_string()))?
-    }
-
-    fn write_bundle_payload(
-        payload: &[u8],
-        plan: BundleWritePlan,
-    ) -> Result<(), AssetExecutionError> {
-        if let Some(path) = plan.raw_target {
-            Self::write_raw_bundle(&path, payload)?;
-        }
-        if let Some(path) = plan.haruki_3d_target {
-            Self::write_haruki_3d_work_bundle(&path, payload)?;
-            Self::record_bundle_payload_hash(
-                plan.bundle_hash_index,
-                plan.bundle_hash_index_key,
-                payload,
-            )?;
-        }
-        if let Some(path) = plan.temp_target {
-            Self::write_temp_bundle(&path, payload)?;
-        }
-        Ok(())
-    }
-
-    fn record_bundle_payload_hash(
-        index: Option<Arc<std::sync::Mutex<DownloadRecord>>>,
-        key: String,
-        payload: &[u8],
-    ) -> Result<(), AssetExecutionError> {
-        let Some(index) = index else {
-            return Ok(());
-        };
-        let digest = hex::encode(Sha256::digest(payload));
-        index
-            .lock()
-            .map_err(|_| {
-                AssetExecutionError::BlockingTask("bundle hash index lock poisoned".to_string())
-            })?
-            .insert(key, digest);
-        Ok(())
-    }
-
-    fn write_temp_bundle(path: &Path, payload: &[u8]) -> Result<(), AssetExecutionError> {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).map_err(|source| {
-                AssetExecutionError::CreateTempDir {
-                    path: parent.to_path_buf(),
-                    source,
-                }
-            })?;
-        }
-        std::fs::write(path, payload).map_err(|source| AssetExecutionError::WriteTempFile {
-            path: path.to_path_buf(),
-            source,
-        })
-    }
-
-    async fn export_bundle_payloads(
-        &self,
-        app_config: &AppConfig,
-        task: &DownloadTask,
-        asset_save_dir: &str,
-        temp_file: &Path,
-        export_path_registry: &NativeSemanticExportPathRegistry,
-    ) -> Result<Option<NativeBundlePostProcessJob>, AssetExecutionError> {
-        let export_started = Instant::now();
-        let payload_export = export_unity_asset_bundle_payloads_with_registry(
-            app_config,
-            &self.region,
-            temp_file,
-            &task.bundle_path,
-            Path::new(asset_save_dir),
-            asset_category_name(&task.category),
-            export_path_registry,
-        )
-        .await;
-        let _ = remove_file_if_exists(temp_file);
-        Ok(Some(NativeBundlePostProcessJob {
-            bundle_path: task.bundle_path.clone(),
-            bundle_hash: task.bundle_hash.clone(),
-            export_started,
-            payload_export: payload_export?,
-            backlog_wait_ms: 0,
-            _backlog_permit: None,
-            _memory_permit: None,
-        }))
-    }
-
-    async fn prefetch_bundle(
-        &self,
-        app_config: &AppConfig,
-        task: &DownloadTask,
-        progress: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-    ) -> Result<(), AssetExecutionError> {
-        let asset_save_dir = self.region.paths.asset_save_dir.clone().ok_or_else(|| {
-            AssetExecutionError::MissingAssetSaveDir {
-                region: self.region_name.clone(),
-            }
-        })?;
-        let bundle_url = self.render_bundle_url(task)?;
-        let download_started = Instant::now();
-        let fetch = self
-            .fetch_deobfuscated_bundle(app_config, &bundle_url, task)
-            .await?;
-        Self::send_progress(
-            progress,
-            ExecutionProgressUpdate::BundleDownloaded {
-                bundle: task.bundle_path.clone(),
-                bytes: fetch.body.len(),
-                elapsed_ms: download_started.elapsed().as_millis(),
-            },
-        );
-        if configured_asset_bundle_cache_dir(app_config).is_none()
-            || self.matches_raw_bundle_filters(&task.bundle_path)
-        {
-            let raw_path = self.raw_bundle_output_path(&asset_save_dir, &task.bundle_path)?;
-            Self::write_raw_bundle(&raw_path, &fetch.body)?;
-            tracing::debug!(
-                region = %self.region_name,
-                bundle = %task.bundle_path,
-                output = %raw_path.display(),
-                http_version = ?app_config.server.asset_http_version,
-                "prefetched raw asset bundle"
-            );
-        }
-        Ok(())
-    }
-
-    async fn fetch_deobfuscated_bundle(
-        &self,
-        app_config: &AppConfig,
-        bundle_url: &str,
-        task: &DownloadTask,
-    ) -> Result<BundleFetch, AssetExecutionError> {
-        let Some(cache_dir) = configured_asset_bundle_cache_dir(app_config) else {
-            let body = self.get_with_retry(bundle_url).await?;
-            return Ok(BundleFetch {
-                body: deobfuscate_owned(body),
-                source: BundleFetchSource::Network,
-            });
-        };
-
-        self.get_bundle_with_cache(bundle_url, task, &cache_dir)
-            .await
-    }
-
-    async fn get_bundle_with_cache(
-        &self,
-        bundle_url: &str,
-        task: &DownloadTask,
-        cache_dir: &Path,
-    ) -> Result<BundleFetch, AssetExecutionError> {
-        let safe_bundle_path = validate_relative_bundle_path(&task.bundle_path)?;
-        let cache_file = cache_dir.join(&self.region_name).join(safe_bundle_path);
-        match Self::bundle_cache_entry_status(&cache_file, task).await {
-            BundleCacheEntryStatus::Current => match tokio::fs::read(&cache_file).await {
-                Ok(body) => {
-                    // Persistent cache entries are already deobfuscated. Returning the owned file
-                    // buffer directly avoids a same-size `data.to_vec()` on every cache hit.
-                    return Ok(BundleFetch {
-                        body,
-                        source: BundleFetchSource::CacheHit,
-                    });
-                }
-                Err(source) if source.kind() == std::io::ErrorKind::NotFound => {
-                    // The cache may have been concurrently replaced between metadata validation
-                    // and the read. Fall back to the network rather than failing the whole job.
-                }
-                Err(source) => {
-                    return Err(AssetExecutionError::ReadTempFile {
-                        path: cache_file,
-                        source,
-                    });
-                }
-            },
-            BundleCacheEntryStatus::Stale => {
-                tracing::warn!(
-                    region = %self.region_name,
-                    bundle = %task.bundle_path,
-                    cache = %cache_file.display(),
-                    "asset bundle cache entry is stale or incomplete; refreshing it"
-                );
-            }
-            BundleCacheEntryStatus::Missing => {}
-        }
-
-        let network_body = self.get_with_retry(bundle_url).await?;
-        let body = deobfuscate_owned(network_body);
-        Self::write_bundle_cache_entry(&cache_file, &task.bundle_hash, &body).await?;
-        Ok(BundleFetch {
-            body,
-            source: BundleFetchSource::CacheMiss,
-        })
-    }
-
-    async fn bundle_cache_entry_status(
-        cache_file: &Path,
-        task: &DownloadTask,
-    ) -> BundleCacheEntryStatus {
-        let metadata_path = bundle_cache_metadata_path(cache_file);
-        match tokio::fs::read_to_string(&metadata_path).await {
-            Ok(cached_hash) => {
-                if cached_hash.trim() != task.bundle_hash {
-                    return if tokio::fs::metadata(cache_file).await.is_ok() {
-                        BundleCacheEntryStatus::Stale
-                    } else {
-                        BundleCacheEntryStatus::Missing
-                    };
-                }
-                match tokio::fs::metadata(cache_file).await {
-                    Ok(_) => BundleCacheEntryStatus::Current,
-                    Err(source) if source.kind() == std::io::ErrorKind::NotFound => {
-                        BundleCacheEntryStatus::Missing
-                    }
-                    Err(_) => BundleCacheEntryStatus::Stale,
-                }
-            }
-            Err(source) if source.kind() == std::io::ErrorKind::NotFound => {
-                // Compatibility with caches produced before hash sidecars were
-                // introduced. Network payloads carry a four-byte obfuscation
-                // marker that the persistent cache omits.
-                let expected = usize::try_from(task.file_size)
-                    .ok()
-                    .filter(|size| *size > 0);
-                match tokio::fs::metadata(cache_file).await {
-                    Ok(metadata) => {
-                        let body_len = usize::try_from(metadata.len()).ok();
-                        if body_len.is_some_and(|body_len| {
-                            expected.is_none_or(|expected| {
-                                body_len == expected || body_len.checked_add(4) == Some(expected)
-                            })
-                        }) {
-                            BundleCacheEntryStatus::Current
-                        } else {
-                            BundleCacheEntryStatus::Stale
-                        }
-                    }
-                    Err(source) if source.kind() == std::io::ErrorKind::NotFound => {
-                        BundleCacheEntryStatus::Missing
-                    }
-                    Err(_) => BundleCacheEntryStatus::Stale,
-                }
-            }
-            Err(_) => BundleCacheEntryStatus::Stale,
-        }
-    }
-
-    async fn write_bundle_cache_entry(
-        cache_file: &Path,
-        bundle_hash: &str,
-        body: &[u8],
-    ) -> Result<(), AssetExecutionError> {
-        if let Some(parent) = cache_file.parent() {
-            tokio::fs::create_dir_all(parent).await.map_err(|source| {
-                AssetExecutionError::CreateTempDir {
-                    path: parent.to_path_buf(),
-                    source,
-                }
-            })?;
-        }
-        Self::atomic_write_bundle_cache_file(cache_file, body).await?;
-        let metadata_path = bundle_cache_metadata_path(cache_file);
-        Self::atomic_write_bundle_cache_file(&metadata_path, bundle_hash.as_bytes()).await
-    }
-
-    async fn atomic_write_bundle_cache_file(
-        path: &Path,
-        body: &[u8],
-    ) -> Result<(), AssetExecutionError> {
-        let sequence = BUNDLE_CACHE_WRITE_SEQUENCE.fetch_add(1, Ordering::Relaxed);
-        let mut temp_name = std::ffi::OsString::from(".");
-        temp_name.push(path.file_name().unwrap_or_default());
-        temp_name.push(format!(".tmp-{}-{sequence}", std::process::id()));
-        let temp_path = path.with_file_name(temp_name);
-        tokio::fs::write(&temp_path, body).await.map_err(|source| {
-            AssetExecutionError::WriteTempFile {
-                path: temp_path.clone(),
-                source,
-            }
-        })?;
-        if let Err(source) = tokio::fs::rename(&temp_path, path).await {
-            let _ = tokio::fs::remove_file(&temp_path).await;
-            return Err(AssetExecutionError::WriteTempFile {
-                path: path.to_path_buf(),
-                source,
-            });
-        }
-        Ok(())
-    }
-
-    fn matches_raw_bundle_filters(&self, bundle_path: &str) -> bool {
-        let Some(raw_bundles) = self.region.export.raw_bundles.as_ref() else {
-            return false;
-        };
-        let include_patterns = compile_patterns(&raw_bundles.include);
-        let exclude_patterns = compile_patterns(&raw_bundles.exclude);
-        (include_patterns.is_empty() || matches_any(&include_patterns, bundle_path))
-            && !matches_any(&exclude_patterns, bundle_path)
-    }
-
-    fn matches_haruki_3d_filters(&self, bundle_path: &str) -> bool {
-        let haruki_3d = &self.region.export.haruki_3d;
-        if !haruki_3d.enabled || haruki_3d.include.is_empty() {
-            return false;
-        }
-        let include_patterns = compile_patterns(&haruki_3d.include);
-        let exclude_patterns = compile_patterns(&haruki_3d.exclude);
-        matches_any(&include_patterns, bundle_path) && !matches_any(&exclude_patterns, bundle_path)
-    }
-
-    fn haruki_3d_work_asset_root(&self) -> Option<PathBuf> {
-        let haruki_3d = &self.region.export.haruki_3d;
-        if !haruki_3d.enabled {
-            return None;
-        }
-        let run_id = self
-            .resolved_asset_version
-            .as_deref()
-            .filter(|value| !value.trim().is_empty())
-            .unwrap_or("current")
-            .replace(['/', '\\', ':'], "_");
-        Some(
-            self.haruki_3d_state_root()?
-                .join(run_id)
-                .join("AssetBundles"),
-        )
-    }
-
-    fn haruki_3d_state_root(&self) -> Option<PathBuf> {
-        let haruki_3d = &self.region.export.haruki_3d;
-        if !haruki_3d.enabled {
-            return None;
-        }
-        Some(Path::new(&Self::haruki_3d_work_dir(haruki_3d)).join(&self.region_name))
-    }
-
-    fn haruki_3d_download_record_path(&self) -> Option<PathBuf> {
-        self.haruki_3d_state_root()
-            .map(|root| root.join("downloaded_assets.json"))
-    }
-
-    fn haruki_3d_bundle_hash_index_path(&self) -> Option<PathBuf> {
-        self.haruki_3d_state_root()
-            .map(|root| root.join("bundle_sha256.json"))
-    }
-
-    fn haruki_3d_bundle_dependency_index_path(&self) -> Option<PathBuf> {
-        self.haruki_3d_state_root()
-            .map(|root| root.join("bundle_dependencies.json"))
-    }
-
-    fn haruki_3d_work_dir(haruki_3d: &crate::core::config::Haruki3dExportConfig) -> String {
-        if !haruki_3d.work_dir.trim().is_empty() {
-            haruki_3d.work_dir.clone()
-        } else {
-            haruki_3d.staging_dir.clone()
-        }
-    }
-
-    fn raw_bundle_output_path(
-        &self,
-        asset_save_dir: &str,
-        bundle_path: &str,
-    ) -> Result<PathBuf, AssetExecutionError> {
-        let root = self
-            .region
-            .export
-            .raw_bundles
-            .as_ref()
-            .and_then(|raw_bundles| raw_bundles.output_dir.as_deref())
-            .map(PathBuf::from)
-            .unwrap_or_else(|| Path::new(asset_save_dir).join("AssetBundles"));
-        raw_bundle_output_path(&root, bundle_path)
-    }
-
-    fn write_raw_bundle(path: &Path, deobfuscated: &[u8]) -> Result<(), AssetExecutionError> {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).map_err(|source| {
-                AssetExecutionError::CreateRawBundleDir {
-                    path: parent.to_path_buf(),
-                    source,
-                }
-            })?;
-        }
-        std::fs::write(path, deobfuscated).map_err(|source| AssetExecutionError::WriteRawBundle {
-            path: path.to_path_buf(),
-            source,
-        })
-    }
-
-    fn write_haruki_3d_work_bundle(
-        path: &Path,
-        deobfuscated: &[u8],
-    ) -> Result<(), AssetExecutionError> {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).map_err(|source| {
-                AssetExecutionError::CreateHaruki3dStagingDir {
-                    path: parent.to_path_buf(),
-                    source,
-                }
-            })?;
-        }
-        std::fs::write(path, deobfuscated).map_err(|source| {
-            AssetExecutionError::WriteHaruki3dStagingBundle {
-                path: path.to_path_buf(),
-                source,
-            }
-        })
-    }
-
-    pub async fn run_haruki_3d_background_export(
-        mut self,
-        _app_config: &AppConfig,
-        progress: Option<UnboundedSender<ExecutionProgressUpdate>>,
-        cancel_flag: Option<Arc<AtomicBool>>,
-    ) -> Result<Haruki3dExportSummary, AssetExecutionError> {
-        if !self.region.export.haruki_3d.enabled {
-            return Ok(Haruki3dExportSummary::default());
-        }
-        self.ensure_not_cancelled(&cancel_flag)?;
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::FetchingAssetInfo,
-                message: "fetching asset bundle info for Haruki 3D export".to_string(),
-            },
-        );
-        let Some(mut plan) = self.prepare_haruki_3d_export_plan().await? else {
-            return Ok(Haruki3dExportSummary::default());
-        };
-        Self::send_progress(
-            &progress,
-            ExecutionProgressUpdate::DownloadsPlanned {
-                total: plan.pending_tasks.len(),
-            },
-        );
-        self.prepare_haruki_3d_staging(&plan, &progress, &cancel_flag)?;
-        self.run_haruki_3d_export_plan(&mut plan, &progress).await?;
-        Self::finish_haruki_3d_export_plan(&plan)?;
-        Ok(Haruki3dExportSummary {
-            matched_bundles: plan.tasks.len(),
-            downloaded_bundles: plan.pending_tasks.len(),
-        })
-    }
-
-    async fn prepare_haruki_3d_export_plan(
-        &mut self,
-    ) -> Result<Option<Haruki3dExportPlan>, AssetExecutionError> {
-        if self.requires_cookies() {
-            self.fetch_runtime_cookies().await?;
-        }
-        let info = self.fetch_asset_bundle_info().await?;
-        let tasks = self.build_haruki_3d_tasks(&info);
-        let dependency_index_path = self.required_haruki_3d_dependency_index_path()?;
-        Self::save_haruki_3d_dependency_index(&dependency_index_path, &info, &tasks)?;
-        let record_path = self.required_haruki_3d_download_record_path()?;
-        let downloaded_assets = load_download_record(&record_path)?;
-        let can_reuse = self.can_reuse_haruki_3d_download_record().await;
-        let pending_tasks = Self::pending_haruki_3d_tasks(&tasks, &downloaded_assets, can_reuse);
-        let pending_paths = pending_tasks
-            .iter()
-            .map(|task| task.bundle_path.clone())
-            .collect();
-        let Some(asset_root) = self.haruki_3d_work_asset_root() else {
-            return Ok(None);
-        };
-        let work_run_dir = asset_root
-            .parent()
-            .map(Path::to_path_buf)
-            .unwrap_or_else(|| asset_root.clone());
-        Ok(Some(Haruki3dExportPlan {
-            config: self.region.export.haruki_3d.clone(),
-            info,
-            tasks,
-            pending_tasks,
-            pending_paths,
-            downloaded_assets,
-            record_path,
-            dependency_index_path,
-            asset_root,
-            work_run_dir,
-        }))
-    }
-
-    fn required_haruki_3d_dependency_index_path(&self) -> Result<PathBuf, AssetExecutionError> {
-        self.haruki_3d_bundle_dependency_index_path()
-            .ok_or_else(|| {
-                AssetExecutionError::BlockingTask(
-                    "3D bundle dependency index path is unavailable".to_string(),
-                )
-            })
-    }
-
-    fn required_haruki_3d_download_record_path(&self) -> Result<PathBuf, AssetExecutionError> {
-        self.haruki_3d_download_record_path().ok_or_else(|| {
-            AssetExecutionError::BlockingTask("3D download record path is unavailable".to_string())
-        })
-    }
-
-    fn pending_haruki_3d_tasks(
-        tasks: &[DownloadTask],
-        downloaded_assets: &DownloadRecord,
-        can_reuse: bool,
-    ) -> Vec<DownloadTask> {
-        tasks
-            .iter()
-            .filter(|task| {
-                !can_reuse
-                    || downloaded_assets
-                        .get(&task.bundle_path)
-                        .is_none_or(|hash| hash != &task.bundle_hash)
-            })
-            .cloned()
-            .collect()
-    }
-
-    fn prepare_haruki_3d_staging(
-        &self,
-        plan: &Haruki3dExportPlan,
-        progress: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-        cancel_flag: &Option<Arc<AtomicBool>>,
-    ) -> Result<(), AssetExecutionError> {
-        self.verify_pending_haruki_3d_bundles(plan, progress, cancel_flag)?;
-        self.create_haruki_3d_sparse_placeholders(plan)?;
-        self.update_haruki_3d_sparse_marker(plan)
-    }
-
-    fn verify_pending_haruki_3d_bundles(
-        &self,
-        plan: &Haruki3dExportPlan,
-        progress: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-        cancel_flag: &Option<Arc<AtomicBool>>,
-    ) -> Result<(), AssetExecutionError> {
-        for task in &plan.pending_tasks {
-            self.ensure_not_cancelled(cancel_flag)?;
-            Self::send_progress(
-                progress,
-                ExecutionProgressUpdate::BundleStarted {
-                    bundle: task.bundle_path.clone(),
-                },
-            );
-            let output_path = raw_bundle_output_path(&plan.asset_root, &task.bundle_path)?;
-            if !output_path.exists() {
-                return Err(AssetExecutionError::MissingHaruki3dStagingBundle {
-                    path: output_path,
-                });
-            }
-            Self::send_progress(
-                progress,
-                ExecutionProgressUpdate::BundleCompleted {
-                    bundle: task.bundle_path.clone(),
-                },
-            );
-        }
-        Ok(())
-    }
-
-    fn create_haruki_3d_sparse_placeholders(
-        &self,
-        plan: &Haruki3dExportPlan,
-    ) -> Result<(), AssetExecutionError> {
-        for task in &plan.tasks {
-            if plan.pending_paths.contains(&task.bundle_path) {
-                continue;
-            }
-            let output_path = raw_bundle_output_path(&plan.asset_root, &task.bundle_path)?;
-            if !output_path.exists() {
-                Self::write_haruki_3d_work_bundle(&output_path, &[])?;
-            }
-        }
-        Ok(())
-    }
-
-    fn update_haruki_3d_sparse_marker(
-        &self,
-        plan: &Haruki3dExportPlan,
-    ) -> Result<(), AssetExecutionError> {
-        let marker = plan.asset_root.join(".haruki-sparse-input");
-        if plan.pending_tasks.len() < plan.tasks.len() {
-            return Self::write_haruki_3d_work_bundle(&marker, &[]);
-        }
-        if marker.exists() {
-            std::fs::remove_file(&marker).map_err(|source| {
-                AssetExecutionError::RemoveHaruki3dStagingDir {
-                    path: marker,
-                    source,
-                }
-            })?;
-        }
-        Ok(())
-    }
-
-    async fn run_haruki_3d_export_plan(
-        &self,
-        plan: &mut Haruki3dExportPlan,
-        progress: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-    ) -> Result<(), AssetExecutionError> {
-        let hash_index_path = self.haruki_3d_bundle_hash_index_path().ok_or_else(|| {
-            AssetExecutionError::BlockingTask(
-                "3D bundle hash index path is unavailable".to_string(),
-            )
-        })?;
-        let commands = Self::build_haruki_3d_exporter_commands(
-            &plan.config,
-            &plan.asset_root,
-            &hash_index_path,
-            &plan.dependency_index_path,
-        );
-        for args in commands {
-            if let Err(error) = self
-                .run_haruki_3d_exporter_stage(&plan.config, &args, progress)
-                .await
-            {
-                self.handle_haruki_3d_export_failure(plan, &error)?;
-                return Err(error);
-            }
-        }
-        let catalog_args = Self::build_haruki_3d_runtime_catalog_command(&plan.config);
-        if let Err(error) = self
-            .run_haruki_3d_exporter_stage(&plan.config, &catalog_args, progress)
-            .await
-        {
-            self.cleanup_failed_haruki_3d_export(plan)?;
-            return Err(error);
-        }
-        Ok(())
-    }
-
-    fn handle_haruki_3d_export_failure(
-        &self,
-        plan: &mut Haruki3dExportPlan,
-        error: &AssetExecutionError,
-    ) -> Result<(), AssetExecutionError> {
-        if let AssetExecutionError::Haruki3dExporterFailed { stderr, .. } = error {
-            self.invalidate_missing_haruki_3d_bundles(plan, stderr)?;
-        }
-        self.cleanup_failed_haruki_3d_export(plan)
-    }
-
-    fn invalidate_missing_haruki_3d_bundles(
-        &self,
-        plan: &mut Haruki3dExportPlan,
-        stderr: &str,
-    ) -> Result<(), AssetExecutionError> {
-        let task_paths: HashSet<_> = plan
-            .tasks
-            .iter()
-            .map(|task| task.bundle_path.as_str())
-            .collect();
-        let recovery_paths = missing_haruki_3d_bundle_paths(stderr)
-            .into_iter()
-            .filter(|path| task_paths.contains(path.as_str()))
-            .flat_map(|path| {
-                std::iter::once(path.clone()).chain(bundle_dependency_closure(&plan.info, &path))
-            })
-            .collect::<HashSet<_>>();
-        let removed = recovery_paths
-            .iter()
-            .filter(|path| plan.downloaded_assets.remove(path.as_str()).is_some())
-            .count();
-        if removed > 0 {
-            save_download_record(&plan.record_path, &plan.downloaded_assets)?;
-            tracing::warn!(
-                region = %self.region_name,
-                removed,
-                "invalidated missing sparse 3D bundles for targeted retry"
-            );
-        }
-        Ok(())
-    }
-
-    fn cleanup_failed_haruki_3d_export(
-        &self,
-        plan: &Haruki3dExportPlan,
-    ) -> Result<(), AssetExecutionError> {
-        if plan.config.cleanup_work_dir_after_failure {
-            Self::remove_haruki_3d_work_dir(&plan.work_run_dir)?;
-        }
-        Ok(())
-    }
-
-    fn finish_haruki_3d_export_plan(plan: &Haruki3dExportPlan) -> Result<(), AssetExecutionError> {
-        let completed_record = plan
-            .tasks
-            .iter()
-            .map(|task| (task.bundle_path.clone(), task.bundle_hash.clone()))
-            .collect();
-        save_download_record(&plan.record_path, &completed_record)?;
-        if plan.config.cleanup_work_dir_after_success {
-            Self::remove_haruki_3d_work_dir(&plan.work_run_dir)?;
-        }
-        Ok(())
-    }
-
-    async fn run_haruki_3d_exporter_stage(
-        &self,
-        haruki_3d: &crate::core::config::Haruki3dExportConfig,
-        args: &[String],
-        progress: &Option<UnboundedSender<ExecutionProgressUpdate>>,
-    ) -> Result<(), AssetExecutionError> {
-        let stage = args.first().map(String::as_str).unwrap_or("unknown");
-        Self::send_progress(
-            progress,
-            ExecutionProgressUpdate::Phase {
-                phase: JobPhase::Exporting3dRuntime,
-                message: format!("running Haruki 3D exporter: {stage}"),
-            },
-        );
-        let exporter_started = Instant::now();
-        let output = tokio::process::Command::new(&haruki_3d.exporter_path)
-            .args(args)
-            .output()
-            .await
-            .map_err(|source| AssetExecutionError::Haruki3dExporterSpawn {
-                program: haruki_3d.exporter_path.clone(),
-                source,
-            })?;
-        if !output.status.success() {
-            return Err(AssetExecutionError::Haruki3dExporterFailed {
-                program: haruki_3d.exporter_path.clone(),
-                status: output.status.to_string(),
-                stderr: String::from_utf8_lossy(&output.stderr).trim().to_string(),
-            });
-        }
-        tracing::info!(
-            region = %self.region_name,
-            %stage,
-            elapsed_ms = exporter_started.elapsed().as_millis(),
-            "Haruki 3D exporter stage completed"
-        );
-        let metrics = exporter_metric_lines(&output.stdout);
-        if !metrics.is_empty() {
-            tracing::info!(region = %self.region_name, %metrics, "Haruki 3D exporter metrics");
-        }
-        tracing::debug!(
-            region = %self.region_name,
-            stdout = %String::from_utf8_lossy(&output.stdout).trim(),
-            stderr = %String::from_utf8_lossy(&output.stderr).trim(),
-            "Haruki 3D exporter stage output"
-        );
-        Ok(())
-    }
-
-    fn build_haruki_3d_runtime_catalog_command(
-        haruki_3d: &crate::core::config::Haruki3dExportConfig,
-    ) -> Vec<String> {
-        vec![
-            "--emit-runtime-role-catalog".to_string(),
-            "--master".to_string(),
-            haruki_3d.master_dir.clone(),
-            "--out".to_string(),
-            haruki_3d.output_dir.clone(),
-        ]
-    }
-
-    fn build_haruki_3d_costume_registry_command(
-        haruki_3d: &crate::core::config::Haruki3dExportConfig,
-        asset_root: &Path,
-    ) -> Vec<String> {
-        vec![
-            "--emit-costume-registries".to_string(),
-            "--master".to_string(),
-            haruki_3d.master_dir.clone(),
-            "--asset-root".to_string(),
-            asset_root.to_string_lossy().into_owned(),
-            "--out".to_string(),
-            haruki_3d.output_dir.clone(),
-            "--convert-model-textures".to_string(),
-            haruki_3d.convert_model_textures.to_string(),
-        ]
-    }
-
-    fn build_haruki_3d_exporter_commands(
-        haruki_3d: &crate::core::config::Haruki3dExportConfig,
-        asset_root: &Path,
-        bundle_hash_index: &Path,
-        bundle_dependency_index: &Path,
-    ) -> Vec<Vec<String>> {
-        let asset_root_arg = asset_root.to_string_lossy().to_string();
-        let model_texture_args = || {
-            vec![
-                "--convert-model-textures".to_string(),
-                haruki_3d.convert_model_textures.to_string(),
-            ]
-        };
-        let mut part_args: Vec<String> = [
-            "--emit-part-packages".to_string(),
-            "--master".to_string(),
-            haruki_3d.master_dir.clone(),
-            "--asset-root".to_string(),
-            asset_root_arg.clone(),
-            "--out".to_string(),
-            haruki_3d.output_dir.clone(),
-            "--manifest".to_string(),
-            haruki_3d.manifest_file.clone(),
-            "--part-package-process-concurrency".to_string(),
-            haruki_3d.process_concurrency.to_string(),
-        ]
-        .into_iter()
-        .chain(model_texture_args())
-        .collect();
-        if !haruki_3d.shared_content_store.trim().is_empty() {
-            part_args.push("--shared-content-store".to_string());
-            part_args.push(haruki_3d.shared_content_store.clone());
-        }
-        if !haruki_3d.compiled_content_store.trim().is_empty() {
-            part_args.push("--compiled-content-store".to_string());
-            part_args.push(haruki_3d.compiled_content_store.clone());
-        }
-        part_args.push("--bundle-hash-index".to_string());
-        part_args.push(bundle_hash_index.to_string_lossy().into_owned());
-        part_args.push("--bundle-dependency-index".to_string());
-        part_args.push(bundle_dependency_index.to_string_lossy().into_owned());
-        let mut role_args = vec![
-            "--emit-role-runtimes".to_string(),
-            "--master".to_string(),
-            haruki_3d.master_dir.clone(),
-            "--asset-root".to_string(),
-            asset_root_arg,
-            "--out".to_string(),
-            haruki_3d.output_dir.clone(),
-        ];
-        role_args.push("--part-package-process-concurrency".to_string());
-        role_args.push(haruki_3d.process_concurrency.to_string());
-        for id in &haruki_3d.role_character3d_ids {
-            role_args.push("--role-character3d-id".to_string());
-            role_args.push(id.to_string());
-        }
-        role_args.extend(model_texture_args());
-        vec![
-            part_args,
-            role_args,
-            Self::build_haruki_3d_costume_registry_command(haruki_3d, asset_root),
-        ]
-    }
-
-    fn build_haruki_3d_tasks(&self, info: &AssetBundleInfo) -> Vec<DownloadTask> {
-        self.build_haruki_3d_filter_tasks(info)
-    }
-
-    fn build_raw_bundle_filter_tasks(&self, info: &AssetBundleInfo) -> Vec<DownloadTask> {
-        let mut tasks = Vec::new();
-        for (bundle_name, detail) in &info.bundles {
-            if !self.matches_raw_bundle_filters(bundle_name) {
-                continue;
-            }
-            let bundle_hash = match self.region.provider {
-                RegionProviderConfig::Nuverse { .. } => detail.crc.to_string(),
-                RegionProviderConfig::ColorfulPalette { .. } => detail.hash.clone(),
-            };
-            tasks.push(DownloadTask {
-                download_path: download_path_for_region(&self.region.provider, bundle_name, detail),
-                bundle_path: bundle_name.clone(),
-                bundle_hash,
-                category: detail.category.clone(),
-                file_size: detail.file_size,
-                priority: usize::MAX,
-                export_payloads: false,
-                stage_haruki_3d: false,
-            });
-        }
-        tasks.sort_by(|a, b| a.bundle_path.cmp(&b.bundle_path));
-        tasks
-    }
-
-    fn build_haruki_3d_download_tasks(
-        &self,
-        info: &AssetBundleInfo,
-        downloaded_assets: &DownloadRecord,
-        can_reuse_download_record: bool,
-    ) -> Result<Vec<DownloadTask>, AssetExecutionError> {
-        if self.haruki_3d_work_asset_root().is_none() {
-            return Ok(Vec::new());
-        }
-        let mut tasks = Vec::new();
-        for task in self.build_haruki_3d_filter_tasks(info) {
-            let has_current_record = can_reuse_download_record
-                && downloaded_assets
-                    .get(&task.bundle_path)
-                    .is_some_and(|existing| existing == &task.bundle_hash);
-            if !has_current_record {
-                tasks.push(DownloadTask {
-                    stage_haruki_3d: true,
-                    ..task
-                });
-            }
-        }
-        Ok(tasks)
-    }
-
-    async fn can_reuse_haruki_3d_download_record(&self) -> bool {
-        let manifest_file = self.region.export.haruki_3d.manifest_file.clone();
-        tokio::task::spawn_blocking(move || {
-            let Ok(bytes) = std::fs::read(manifest_file) else {
-                return false;
-            };
-            sonic_rs::from_slice::<HashMap<String, sonic_rs::Value>>(&bytes)
-                .is_ok_and(|entries| !entries.is_empty())
-        })
-        .await
-        .unwrap_or(false)
-    }
-
-    fn build_haruki_3d_filter_tasks(&self, info: &AssetBundleInfo) -> Vec<DownloadTask> {
-        let mut selected: HashSet<String> = info
-            .bundles
-            .keys()
-            .filter(|bundle_name| self.matches_haruki_3d_filters(bundle_name))
-            .cloned()
-            .collect();
-        let mut pending: Vec<String> = selected.iter().cloned().collect();
-        while let Some(bundle_name) = pending.pop() {
-            let Some(detail) = info.bundles.get(&bundle_name) else {
-                continue;
-            };
-            for dependency in &detail.dependencies {
-                if info.bundles.contains_key(dependency) && selected.insert(dependency.clone()) {
-                    pending.push(dependency.clone());
-                }
-            }
-        }
-        let mut tasks = Vec::new();
-        for (bundle_name, detail) in &info.bundles {
-            if !selected.contains(bundle_name) {
-                continue;
-            }
-            let bundle_hash = match self.region.provider {
-                RegionProviderConfig::Nuverse { .. } => detail.crc.to_string(),
-                RegionProviderConfig::ColorfulPalette { .. } => detail.hash.clone(),
-            };
-            tasks.push(DownloadTask {
-                download_path: download_path_for_region(&self.region.provider, bundle_name, detail),
-                bundle_path: bundle_name.clone(),
-                bundle_hash,
-                category: detail.category.clone(),
-                file_size: detail.file_size,
-                priority: usize::MAX,
-                export_payloads: false,
-                stage_haruki_3d: true,
-            });
-        }
-        tasks.sort_by(|a, b| a.bundle_path.cmp(&b.bundle_path));
-        tasks
-    }
-
-    fn save_haruki_3d_dependency_index(
-        path: &Path,
-        info: &AssetBundleInfo,
-        tasks: &[DownloadTask],
-    ) -> Result<(), AssetExecutionError> {
-        let index: HashMap<String, Vec<String>> = tasks
-            .iter()
-            .map(|task| {
-                (
-                    task.bundle_path.clone(),
-                    bundle_dependency_closure(info, &task.bundle_path),
-                )
-            })
-            .collect();
-        let bytes = sonic_rs::to_vec_pretty(&index)
-            .map_err(|source| AssetExecutionError::BlockingTask(source.to_string()))?;
-        Self::write_haruki_3d_work_bundle(path, &bytes)
-    }
-
-    fn remove_haruki_3d_work_dir(work_run_dir: &Path) -> Result<(), AssetExecutionError> {
-        if !work_run_dir.exists() {
-            return Ok(());
-        }
-        std::fs::remove_dir_all(work_run_dir).map_err(|source| {
-            AssetExecutionError::RemoveHaruki3dStagingDir {
-                path: work_run_dir.to_path_buf(),
-                source,
-            }
-        })
-    }
-}
-
-/// Validate an untrusted, server-provided bundle path: it must be a relative path made only of
-/// normal components (no empty / `.` / `..` / absolute / root / prefix). Returns it as a relative
-/// `Path` so callers can safely `join` it onto a trusted root without escaping it.
-fn validate_relative_bundle_path(bundle_path: &str) -> Result<&Path, AssetExecutionError> {
-    let invalid = |reason: &str| AssetExecutionError::InvalidRawBundlePath {
-        bundle: bundle_path.to_string(),
-        reason: reason.to_string(),
-    };
-    if bundle_path.is_empty() {
-        return Err(invalid("path is empty"));
-    }
-    if bundle_path
-        .split('/')
-        .any(|component| component.is_empty() || component == "." || component == "..")
-    {
-        return Err(invalid(
-            "empty, current-directory, or parent-directory components are not allowed",
-        ));
-    }
-
-    let relative = Path::new(bundle_path);
-    if relative.is_absolute() {
-        return Err(invalid("absolute paths are not allowed"));
-    }
-
-    for component in relative.components() {
-        match component {
-            Component::Normal(_) => {}
-            Component::CurDir => {
-                return Err(invalid("current-directory components are not allowed"))
-            }
-            Component::ParentDir => {
-                return Err(invalid("parent-directory components are not allowed"))
-            }
-            Component::RootDir | Component::Prefix(_) => {
-                return Err(invalid("root or prefix components are not allowed"))
-            }
-        }
-    }
-
-    Ok(relative)
-}
-
-fn raw_bundle_output_path(root: &Path, bundle_path: &str) -> Result<PathBuf, AssetExecutionError> {
-    let relative = validate_relative_bundle_path(bundle_path)?;
-
-    let mut path = root.to_path_buf();
-    for component in relative.components() {
-        if let Component::Normal(value) = component {
-            path.push(value);
-        }
-    }
-
-    if path.extension().and_then(|ext| ext.to_str()) != Some("bundle") {
-        path.set_extension("bundle");
-    }
-    Ok(path)
-}
-
-#[derive(Debug)]
-struct BundleFetch {
-    /// Deobfuscated bytes ready to pass directly to unity-rs.
-    body: Vec<u8>,
-    #[cfg_attr(not(test), allow(dead_code))]
-    source: BundleFetchSource,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum BundleCacheEntryStatus {
-    Current,
-    Stale,
-    Missing,
-}
-
-#[derive(Debug)]
-enum BundleFetchSource {
-    CacheHit,
-    CacheMiss,
-    Network,
-}
-
-#[cfg(test)]
-impl BundleFetchSource {
-    const fn as_str(&self) -> &'static str {
-        match self {
-            Self::CacheHit => "cache_hit",
-            Self::CacheMiss => "cache_miss",
-            Self::Network => "network",
-        }
-    }
-}
-
-fn configured_asset_bundle_cache_dir(app_config: &AppConfig) -> Option<PathBuf> {
-    app_config
-        .execution
-        .asset_bundle_cache_dir
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(PathBuf::from)
-}
-
-fn bundle_cache_metadata_path(cache_file: &Path) -> PathBuf {
-    let mut file_name = std::ffi::OsString::from(".");
-    file_name.push(cache_file.file_name().unwrap_or_default());
-    file_name.push(".haruki-cache-hash");
-    cache_file.with_file_name(file_name)
-}
-
-fn bundle_hash_index_key(bundle_path: &str) -> Result<String, AssetExecutionError> {
-    Ok(raw_bundle_output_path(Path::new(""), bundle_path)?
-        .to_string_lossy()
-        .replace('\\', "/"))
-}
-
-fn bundle_dependency_closure(info: &AssetBundleInfo, bundle_name: &str) -> Vec<String> {
-    let mut closure = HashSet::new();
-    let mut pending = vec![bundle_name.to_string()];
-    while let Some(current) = pending.pop() {
-        let Some(detail) = info.bundles.get(&current) else {
-            continue;
-        };
-        for dependency in &detail.dependencies {
-            if closure.insert(dependency.clone()) {
-                pending.push(dependency.clone());
-            }
-        }
-    }
-    closure.remove(bundle_name);
-    let mut result: Vec<_> = closure.into_iter().collect();
-    result.sort();
-    result
-}
-
-fn exporter_metric_lines(stdout: &[u8]) -> String {
-    String::from_utf8_lossy(stdout)
-        .lines()
-        .filter(|line| line.contains(" metrics:") || line.starts_with("Planned "))
-        .collect::<Vec<_>>()
-        .join(" | ")
-}
-
-fn missing_haruki_3d_bundle_paths(stderr: &str) -> Vec<String> {
-    const PREFIX: &str = "HARUKI_3D_MISSING_BUNDLE=";
-    let mut paths: Vec<_> = stderr
-        .lines()
-        .filter_map(|line| line.trim().strip_prefix(PREFIX))
-        .filter(|path| !path.is_empty() && raw_bundle_output_path(Path::new(""), path).is_ok())
-        .map(str::to_string)
-        .collect();
-    paths.sort();
-    paths.dedup();
-    paths
-}
-
-#[derive(Clone)]
-struct BundleMemoryLimiter {
-    semaphore: Option<Arc<Semaphore>>,
-    limit_bytes: usize,
-    limit_units: u32,
-}
-
-impl BundleMemoryLimiter {
-    const UNIT_BYTES: usize = 1024 * 1024;
-
-    fn from_config(app_config: &AppConfig) -> Self {
-        let limit_bytes = app_config.resources.memory.max_in_flight_bundle_bytes;
-        if limit_bytes == 0 {
-            return Self {
-                semaphore: None,
-                limit_bytes,
-                limit_units: 0,
-            };
-        }
-        let limit_units = bytes_to_units(limit_bytes).min(u32::MAX as usize).max(1) as u32;
-        Self {
-            semaphore: Some(Arc::new(Semaphore::new(limit_units as usize))),
-            limit_bytes,
-            limit_units,
-        }
-    }
-
-    fn limit_bytes(&self) -> usize {
-        self.limit_bytes
-    }
-
-    async fn acquire(&self, estimated_bytes: usize) -> Option<OwnedSemaphorePermit> {
-        let semaphore = self.semaphore.as_ref()?;
-        let units = bytes_to_units(estimated_bytes)
-            .min(self.limit_units as usize)
-            .max(1) as u32;
-        semaphore.clone().acquire_many_owned(units).await.ok()
-    }
-}
-
-fn bytes_to_units(bytes: usize) -> usize {
-    bytes.div_ceil(BundleMemoryLimiter::UNIT_BYTES).max(1)
-}
-
-fn blocking_panic_message(panic: Box<dyn std::any::Any + Send>) -> String {
-    if let Some(message) = panic.downcast_ref::<&str>() {
-        (*message).to_string()
-    } else if let Some(message) = panic.downcast_ref::<String>() {
-        message.clone()
-    } else {
-        "blocking task panicked".to_string()
-    }
-}
-
-pub async fn fetch_live_asset_bundle_info(
-    app_config: &AppConfig,
-    region_name: &str,
-    region: &RegionConfig,
-    request: &AssetUpdateRequest,
-) -> Result<AssetBundleInfo, AssetExecutionError> {
-    let mut context = AssetExecutionContext::new(app_config, region_name, region, request)?;
-    if context.requires_cookies() {
-        context.fetch_runtime_cookies().await?;
-    }
-    context.fetch_asset_bundle_info().await
-}
-
-fn is_retryable_http_error(err: &AssetExecutionError) -> bool {
-    match err {
-        AssetExecutionError::Http(_) => true,
-        // 5xx are transient; 429 (Too Many Requests) and 408 (Request Timeout) are the canonical
-        // "back off and retry" signals that Project Sekai CDNs/rate limiters emit under load.
-        AssetExecutionError::HttpStatus { status, .. } => {
-            *status >= 500 || *status == 429 || *status == 408
-        }
-        _ => false,
-    }
-}
-
-pub fn decrypt_asset_bundle_info(
-    aes_key_hex: &str,
-    aes_iv_hex: &str,
-    content: &[u8],
-) -> Result<AssetBundleInfo, AssetExecutionError> {
-    if content.is_empty() {
-        return Err(AssetExecutionError::EmptyEncryptedContent);
-    }
-    if !content.len().is_multiple_of(16) {
-        return Err(AssetExecutionError::InvalidEncryptedBlockSize);
-    }
-
-    let key = hex::decode(aes_key_hex)
-        .map_err(|err| AssetExecutionError::InvalidAesKeyHex(err.to_string()))?;
-    let iv = hex::decode(aes_iv_hex)
-        .map_err(|err| AssetExecutionError::InvalidAesIvHex(err.to_string()))?;
-    if iv.len() != 16 {
-        return Err(AssetExecutionError::InvalidAesIvLength { got: iv.len() });
-    }
-
-    let mut buf = content.to_vec();
-    let decrypted = match key.len() {
-        16 => Aes128CbcDec::new_from_slices(&key, &iv)
-            .map_err(|err| AssetExecutionError::AssetInfoDecode(err.to_string()))?
-            .decrypt_padded::<Pkcs7>(&mut buf)
-            .map_err(|err| AssetExecutionError::AssetInfoDecode(err.to_string()))?,
-        24 => Aes192CbcDec::new_from_slices(&key, &iv)
-            .map_err(|err| AssetExecutionError::AssetInfoDecode(err.to_string()))?
-            .decrypt_padded::<Pkcs7>(&mut buf)
-            .map_err(|err| AssetExecutionError::AssetInfoDecode(err.to_string()))?,
-        32 => Aes256CbcDec::new_from_slices(&key, &iv)
-            .map_err(|err| AssetExecutionError::AssetInfoDecode(err.to_string()))?
-            .decrypt_padded::<Pkcs7>(&mut buf)
-            .map_err(|err| AssetExecutionError::AssetInfoDecode(err.to_string()))?,
-        _ => {
-            return Err(AssetExecutionError::AssetInfoDecode(format!(
-                "unsupported AES key length {}",
-                key.len()
-            )))
-        }
-    };
-
-    rmp_serde::from_slice::<AssetBundleInfo>(decrypted)
-        .map_err(|err| AssetExecutionError::AssetInfoDecode(err.to_string()))
-}
-
-pub fn deobfuscate(data: &[u8]) -> Vec<u8> {
-    deobfuscate_owned(data.to_vec())
-}
-
-fn deobfuscate_owned(mut data: Vec<u8>) -> Vec<u8> {
-    const SIMPLE: [u8; 4] = [0x20, 0x00, 0x00, 0x00];
-    const XOR_HEADER: [u8; 4] = [0x10, 0x00, 0x00, 0x00];
-
-    if data.starts_with(&SIMPLE) {
-        data.copy_within(4.., 0);
-        data.truncate(data.len() - 4);
-        return data;
-    }
-
-    if data.starts_with(&XOR_HEADER) {
-        data.copy_within(4.., 0);
-        data.truncate(data.len() - 4);
-        if data.len() < 128 {
-            return data;
-        }
-        let pattern = [0xff, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00];
-        for idx in 0..128 {
-            data[idx] ^= pattern[idx % pattern.len()];
-        }
-        return data;
-    }
-
-    data
-}
-
-pub fn should_download_bundle(
-    region: &RegionConfig,
-    bundle_name: &str,
-    category: &AssetCategory,
-) -> bool {
-    let compiled = match category {
-        AssetCategory::StartApp => compile_patterns(&region.filters.start_app),
-        AssetCategory::OnDemand | AssetCategory::LivePv => {
-            compile_patterns(&region.filters.on_demand)
-        }
-        AssetCategory::Other(_) => return false,
-    };
-    if compiled.is_empty() {
-        return false;
-    }
-    matches_any(&compiled, bundle_name)
-}
-
-fn post_process_backlog_capacity(
-    download_concurrency: usize,
-    post_process_concurrency: usize,
-) -> usize {
-    let _ = download_concurrency;
-    post_process_concurrency.saturating_mul(2).max(1)
-}
-
-fn download_path_for_region(
-    provider: &RegionProviderConfig,
-    bundle_name: &str,
-    detail: &AssetBundleDetail,
-) -> String {
-    match provider {
-        RegionProviderConfig::ColorfulPalette { .. } => bundle_name.to_string(),
-        RegionProviderConfig::Nuverse { .. } => detail
-            .download_path
-            .as_ref()
-            .map(|prefix| format!("{prefix}/{bundle_name}"))
-            .unwrap_or_else(|| bundle_name.to_string()),
-    }
-}
-
-fn time_arg_jst() -> String {
-    let tz = FixedOffset::east_opt(9 * 3600).unwrap();
-    format!(
-        "?t={}",
-        chrono::Utc::now().with_timezone(&tz).format("%Y%m%d%H%M%S")
-    )
-}
+pub use planning::should_download_bundle;
+pub use progress::ExecutionProgressUpdate;
+pub use provider::fetch_live_asset_bundle_info;
 
 #[cfg(test)]
 mod tests {
@@ -2995,13 +43,17 @@ mod tests {
     use crate::core::download_records::DownloadRecord;
     use crate::core::models::{AssetUpdateMode, AssetUpdateRequest};
 
-    use super::{
-        bundle_cache_metadata_path, bundle_dependency_closure, bundle_hash_index_key,
-        decrypt_asset_bundle_info, deobfuscate, deobfuscate_owned, exporter_metric_lines,
-        missing_haruki_3d_bundle_paths, post_process_backlog_capacity, raw_bundle_output_path,
-        should_download_bundle, AssetBundleDetail, AssetBundleInfo, AssetCategory,
-        AssetExecutionContext, BundleCacheEntryStatus, DownloadTask,
+    use super::cache::{bundle_cache_metadata_path, bundle_hash_index_key};
+    use super::crypto::{decrypt_asset_bundle_info, deobfuscate, deobfuscate_owned};
+    use super::haruki_3d::{
+        bundle_dependency_closure, exporter_metric_lines, missing_haruki_3d_bundle_paths,
     };
+    use super::model::{
+        AssetBundleDetail, AssetBundleInfo, AssetCategory, AssetExecutionContext,
+        BundleCacheEntryStatus, DownloadTask,
+    };
+    use super::planning::{raw_bundle_output_path, should_download_bundle};
+    use super::runner::post_process_backlog_capacity;
 
     type Aes128CbcEnc = cbc::Encryptor<aes::Aes128>;
     const TEST_AES_KEY_HEX: &str = "00112233445566778899aabbccddeeff";
